refactor: proto 수정에 맞게 entity 변경
