+++ src/gen/com/hearlers/v1/model/counselor_pb.ts
@@ -1,0 +1,171 @@
+// @generated by protoc-gen-es v2.2.1 with parameter "target=ts"
+// @generated from file com/hearlers/v1/model/counselor.proto (package com.hearlers.v1.model, syntax proto3)
+/* eslint-disable */
+
+import type { GenEnum, GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
+import { enumDesc, fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
+import type { Message } from "@bufbuild/protobuf";
+
+/**
+ * Describes the file com/hearlers/v1/model/counselor.proto.
+ */
+export const file_com_hearlers_v1_model_counselor: GenFile = /*@__PURE__*/
+  fileDesc("CiVjb20vaGVhcmxlcnMvdjEvbW9kZWwvY291bnNlbG9yLnByb3RvEhVjb20uaGVhcmxlcnMudjEubW9kZWwizwIKCUNvdW5zZWxvchIKCgJpZBgBIAEoCRIPCgd0b25lX2lkGAIgASgJEgwKBG5hbWUYAyABKAkSEwoLZGVzY3JpcHRpb24YBCABKAkSNgoGZ2VuZGVyGAUgASgOMiYuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxvckdlbmRlchIvCgdwZXJzb25hGAYgASgLMh4uY29tLmhlYXJsZXJzLnYxLm1vZGVsLlBlcnNvbmESFQoNaW50cm9fbWVzc2FnZRgHIAEoCRIYChByZXNwb25zZV9vcHRpb24xGAggASgJEhgKEHJlc3BvbnNlX29wdGlvbjIYCSABKAkSEgoKY3JlYXRlZF9hdBgKIAEoCRISCgp1cGRhdGVkX2F0GAsgASgJEhcKCmRlbGV0ZWRfYXQYDCABKAlIAIgBAUINCgtfZGVsZXRlZF9hdCKJAQoHUGVyc29uYRIKCgJpZBgBIAEoCRIMCgRib2R5GAIgASgJEhQKDGNvdW5zZWxvcl9pZBgDIAEoCRISCgpjcmVhdGVkX2F0GAQgASgJEhIKCnVwZGF0ZWRfYXQYBSABKAkSFwoKZGVsZXRlZF9hdBgGIAEoCUgAiAEBQg0KC19kZWxldGVkX2F0KoYBCg9Db3Vuc2Vsb3JHZW5kZXISIAocQ09VTlNFTE9SX0dFTkRFUl9VTlNQRUNJRklFRBAAEhkKFUNPVU5TRUxPUl9HRU5ERVJfTUFMRRABEhsKF0NPVU5TRUxPUl9HRU5ERVJfRkVNQUxFEAISGQoVQ09VTlNFTE9SX0dFTkRFUl9OT05FEANCowEKGWNvbS5jb20uaGVhcmxlcnMudjEubW9kZWxCDkNvdW5zZWxvclByb3RvUAGiAgRDSFZNqgIVQ29tLkhlYXJsZXJzLlYxLk1vZGVsygIVQ29tXEhlYXJsZXJzXFYxXE1vZGVs4gIhQ29tXEhlYXJsZXJzXFYxXE1vZGVsXEdQQk1ldGFkYXRh6gIYQ29tOjpIZWFybGVyczo6VjE6Ok1vZGVsYgZwcm90bzM");
+
+/**
+ * @generated from message com.hearlers.v1.model.Counselor
+ */
+export type Counselor = Message<"com.hearlers.v1.model.Counselor"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string tone_id = 2;
+   */
+  toneId: string;
+
+  /**
+   * @generated from field: string name = 3;
+   */
+  name: string;
+
+  /**
+   * @generated from field: string description = 4;
+   */
+  description: string;
+
+  /**
+   * @generated from field: com.hearlers.v1.model.CounselorGender gender = 5;
+   */
+  gender: CounselorGender;
+
+  /**
+   * @generated from field: com.hearlers.v1.model.Persona persona = 6;
+   */
+  persona?: Persona;
+
+  /**
+   * @generated from field: string intro_message = 7;
+   */
+  introMessage: string;
+
+  /**
+   * @generated from field: string response_option1 = 8;
+   */
+  responseOption1: string;
+
+  /**
+   * @generated from field: string response_option2 = 9;
+   */
+  responseOption2: string;
+
+  /**
+   * ISO 8601 (2024-12-29T12:34:56.000Z)
+   *
+   * @generated from field: string created_at = 10;
+   */
+  createdAt: string;
+
+  /**
+   * ISO 8601 (2024-12-29T12:34:56.000Z)
+   *
+   * @generated from field: string updated_at = 11;
+   */
+  updatedAt: string;
+
+  /**
+   * ISO 8601 (2024-12-29T12:34:56.000Z)
+   *
+   * @generated from field: optional string deleted_at = 12;
+   */
+  deletedAt?: string;
+};
+
+/**
+ * Describes the message com.hearlers.v1.model.Counselor.
+ * Use `create(CounselorSchema)` to create a new message.
+ */
+export const CounselorSchema: GenMessage<Counselor> = /*@__PURE__*/
+  messageDesc(file_com_hearlers_v1_model_counselor, 0);
+
+/**
+ * @generated from message com.hearlers.v1.model.Persona
+ */
+export type Persona = Message<"com.hearlers.v1.model.Persona"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string body = 2;
+   */
+  body: string;
+
+  /**
+   * @generated from field: string counselor_id = 3;
+   */
+  counselorId: string;
+
+  /**
+   * ISO 8601 (2024-12-29T12:34:56.000Z)
+   *
+   * @generated from field: string created_at = 4;
+   */
+  createdAt: string;
+
+  /**
+   * ISO 8601 (2024-12-29T12:34:56.000Z)
+   *
+   * @generated from field: string updated_at = 5;
+   */
+  updatedAt: string;
+
+  /**
+   * ISO 8601 (2024-12-29T12:34:56.000Z)
+   *
+   * @generated from field: optional string deleted_at = 6;
+   */
+  deletedAt?: string;
+};
+
+/**
+ * Describes the message com.hearlers.v1.model.Persona.
+ * Use `create(PersonaSchema)` to create a new message.
+ */
+export const PersonaSchema: GenMessage<Persona> = /*@__PURE__*/
+  messageDesc(file_com_hearlers_v1_model_counselor, 1);
+
+/**
+ * @generated from enum com.hearlers.v1.model.CounselorGender
+ */
+export enum CounselorGender {
+  /**
+   * @generated from enum value: COUNSELOR_GENDER_UNSPECIFIED = 0;
+   */
+  UNSPECIFIED = 0,
+
+  /**
+   * @generated from enum value: COUNSELOR_GENDER_MALE = 1;
+   */
+  MALE = 1,
+
+  /**
+   * @generated from enum value: COUNSELOR_GENDER_FEMALE = 2;
+   */
+  FEMALE = 2,
+
+  /**
+   * @generated from enum value: COUNSELOR_GENDER_NONE = 3;
+   */
+  NONE = 3,
+}
+
+/**
+ * Describes the enum com.hearlers.v1.model.CounselorGender.
+ */
+export const CounselorGenderSchema: GenEnum<CounselorGender> = /*@__PURE__*/
+  enumDesc(file_com_hearlers_v1_model_counselor, 0);
+
