+++ src/gen/com/hearlers/v1/service/counsel_pb.ts
@@ -4,7 +4,7 @@
 
 import type { GenFile, GenMessage, GenService } from "@bufbuild/protobuf/codegenv1";
 import { fileDesc, messageDesc, serviceDesc } from "@bufbuild/protobuf/codegenv1";
-import type { Context, Counsel, CounselMessage, CounselMessageReaction, Counselor, CounselorGender, CounselorUserRelationship, CounselTechnique, CounselTechniqueStage, Instruction, InstructionItem, Persona, Tone } from "../model/counsel_pb";
+import type { Counsel, CounselMessage, CounselMessageReaction, CounselorUserRelationship } from "../model/counsel_pb";
 import { file_com_hearlers_v1_model_counsel } from "../model/counsel_pb";
 import type { Message } from "@bufbuild/protobuf";
 
@@ -12,7 +12,7 @@ import type { Message } from "@bufbuild/protobuf";
  * Describes the file com/hearlers/v1/service/counsel.proto.
  */
 export const file_com_hearlers_v1_service_counsel: GenFile = /*@__PURE__*/
-  fileDesc("CiVjb20vaGVhcmxlcnMvdjEvc2VydmljZS9jb3Vuc2VsLnByb3RvEhdjb20uaGVhcmxlcnMudjEuc2VydmljZSKfAQoUQ3JlYXRlQ291bnNlbFJlcXVlc3QSDwoHdXNlcl9pZBgBIAEoCRIUCgxjb3Vuc2Vsb3JfaWQYAiABKAkSGgoNaW50cm9fbWVzc2FnZRgDIAEoCUgAiAEBEh0KEHJlc3BvbnNlX21lc3NhZ2UYBCABKAlIAYgBAUIQCg5faW50cm9fbWVzc2FnZUITChFfcmVzcG9uc2VfbWVzc2FnZSKJAQoVQ3JlYXRlQ291bnNlbFJlc3BvbnNlEi8KB2NvdW5zZWwYASABKAsyHi5jb20uaGVhcmxlcnMudjEubW9kZWwuQ291bnNlbBI/ChBjb3Vuc2VsX21lc3NhZ2VzGAIgAygLMiUuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxNZXNzYWdlIlIKE0ZpbmRDb3Vuc2Vsc1JlcXVlc3QSDwoHdXNlcl9pZBgBIAEoCRIZCgxjb3Vuc2Vsb3JfaWQYAiABKAlIAIgBAUIPCg1fY291bnNlbG9yX2lkIkgKFEZpbmRDb3Vuc2Vsc1Jlc3BvbnNlEjAKCGNvdW5zZWxzGAEgAygLMh4uY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWwiLAoWRmluZENvdW5zZWxCeUlkUmVxdWVzdBISCgpjb3Vuc2VsX2lkGAEgASgJIlsKF0ZpbmRDb3Vuc2VsQnlJZFJlc3BvbnNlEjQKB2NvdW5zZWwYASABKAsyHi5jb20uaGVhcmxlcnMudjEubW9kZWwuQ291bnNlbEgAiAEBQgoKCF9jb3Vuc2VsIjsKFENyZWF0ZU1lc3NhZ2VSZXF1ZXN0EhIKCmNvdW5zZWxfaWQYASABKAkSDwoHbWVzc2FnZRgCIAEoCSKqAQoVQ3JlYXRlTWVzc2FnZVJlc3BvbnNlEkYKF2NyZWF0ZWRfY291bnNlbF9tZXNzYWdlGAEgASgLMiUuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxNZXNzYWdlEkkKGmNvdW5zZWxvcl9yZXNwb25zZV9tZXNzYWdlGAIgASgLMiUuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxNZXNzYWdlIikKE0ZpbmRNZXNzYWdlc1JlcXVlc3QSEgoKY291bnNlbF9pZBgBIAEoCSJXChRGaW5kTWVzc2FnZXNSZXNwb25zZRI/ChBjb3Vuc2VsX21lc3NhZ2VzGAEgAygLMiUuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxNZXNzYWdlImoKE1JlYWN0TWVzc2FnZVJlcXVlc3QSEgoKbWVzc2FnZV9pZBgBIAEoCRI/CghyZWFjdGlvbhgCIAEoDjItLmNvbS5oZWFybGVycy52MS5tb2RlbC5Db3Vuc2VsTWVzc2FnZVJlYWN0aW9uIlYKFFJlYWN0TWVzc2FnZVJlc3BvbnNlEj4KD2NvdW5zZWxfbWVzc2FnZRgBIAEoCzIlLmNvbS5oZWFybGVycy52MS5tb2RlbC5Db3Vuc2VsTWVzc2FnZSKOAQoWQ3JlYXRlQ291bnNlbG9yUmVxdWVzdBIPCgd0b25lX2lkGAEgASgJEgwKBG5hbWUYAiABKAkSEwoLZGVzY3JpcHRpb24YAyABKAkSQAoQY291bnNlbG9yX2dlbmRlchgEIAEoDjImLmNvbS5oZWFybGVycy52MS5tb2RlbC5Db3Vuc2Vsb3JHZW5kZXIiTgoXQ3JlYXRlQ291bnNlbG9yUmVzcG9uc2USMwoJY291bnNlbG9yGAEgASgLMiAuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxvciI5ChVGaW5kQ291bnNlbG9yc1JlcXVlc3QSFAoHdG9uZV9pZBgBIAEoCUgAiAEBQgoKCF90b25lX2lkIk4KFkZpbmRDb3Vuc2Vsb3JzUmVzcG9uc2USNAoKY291bnNlbG9ycxgBIAMoCzIgLmNvbS5oZWFybGVycy52MS5tb2RlbC5Db3Vuc2Vsb3IiMAoYRmluZENvdW5zZWxvckJ5SWRSZXF1ZXN0EhQKDGNvdW5zZWxvcl9pZBgBIAEoCSJjChlGaW5kQ291bnNlbG9yQnlJZFJlc3BvbnNlEjgKCWNvdW5zZWxvchgBIAEoCzIgLmNvbS5oZWFybGVycy52MS5tb2RlbC5Db3Vuc2Vsb3JIAIgBAUIMCgpfY291bnNlbG9yIvIBChZVcGRhdGVDb3Vuc2Vsb3JSZXF1ZXN0EhQKDGNvdW5zZWxvcl9pZBgBIAEoCRIUCgd0b25lX2lkGAIgASgJSACIAQESEQoEbmFtZRgDIAEoCUgBiAEBEhgKC2Rlc2NyaXB0aW9uGAQgASgJSAKIAQESRQoQY291bnNlbG9yX2dlbmRlchgFIAEoDjImLmNvbS5oZWFybGVycy52MS5tb2RlbC5Db3Vuc2Vsb3JHZW5kZXJIA4gBAUIKCghfdG9uZV9pZEIHCgVfbmFtZUIOCgxfZGVzY3JpcHRpb25CEwoRX2NvdW5zZWxvcl9nZW5kZXIiTgoXVXBkYXRlQ291bnNlbG9yUmVzcG9uc2USMwoJY291bnNlbG9yGAEgASgLMiAuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxvciJOCiVGaW5kQ291bnNlbG9yVXNlclJlbGF0aW9uc2hpcHNSZXF1ZXN0EhQKDGNvdW5zZWxvcl9pZBgBIAEoCRIPCgd1c2VyX2lkGAIgASgJIoABCiZGaW5kQ291bnNlbG9yVXNlclJlbGF0aW9uc2hpcHNSZXNwb25zZRJWChxjb3Vuc2Vsb3JfdXNlcl9yZWxhdGlvbnNoaXBzGAEgAygLMjAuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxvclVzZXJSZWxhdGlvbnNoaXAiygEKHUNyZWF0ZUNvdW5zZWxUZWNobmlxdWVSZXF1ZXN0EgwKBG5hbWUYASABKAkSFAoHdG9uZV9pZBgCIAEoCUgAiAEBEhIKCmNvbnRleHRfaWQYAyABKAkSFgoOaW5zdHJ1Y3Rpb25faWQYBCABKAkSTQoXY291bnNlbF90ZWNobmlxdWVfc3RhZ2UYBSABKA4yLC5jb20uaGVhcmxlcnMudjEubW9kZWwuQ291bnNlbFRlY2huaXF1ZVN0YWdlQgoKCF90b25lX2lkImQKHkNyZWF0ZUNvdW5zZWxUZWNobmlxdWVSZXNwb25zZRJCChFjb3Vuc2VsX3RlY2huaXF1ZRgBIAEoCzInLmNvbS5oZWFybGVycy52MS5tb2RlbC5Db3Vuc2VsVGVjaG5pcXVlIswBChxGaW5kQ291bnNlbFRlY2huaXF1ZXNSZXF1ZXN0EhEKBG5hbWUYASABKAlIAIgBARIUCgd0b25lX2lkGAIgASgJSAGIAQESUgoXY291bnNlbF90ZWNobmlxdWVfc3RhZ2UYAyABKA4yLC5jb20uaGVhcmxlcnMudjEubW9kZWwuQ291bnNlbFRlY2huaXF1ZVN0YWdlSAKIAQFCBwoFX25hbWVCCgoIX3RvbmVfaWRCGgoYX2NvdW5zZWxfdGVjaG5pcXVlX3N0YWdlImQKHUZpbmRDb3Vuc2VsVGVjaG5pcXVlc1Jlc3BvbnNlEkMKEmNvdW5zZWxfdGVjaG5pcXVlcxgBIAMoCzInLmNvbS5oZWFybGVycy52MS5tb2RlbC5Db3Vuc2VsVGVjaG5pcXVlIj8KH0ZpbmRDb3Vuc2VsVGVjaG5pcXVlQnlJZFJlcXVlc3QSHAoUY291bnNlbF90ZWNobmlxdWVfaWQYASABKAkigQEKIEZpbmRDb3Vuc2VsVGVjaG5pcXVlQnlJZFJlc3BvbnNlEkcKEWNvdW5zZWxfdGVjaG5pcXVlGAEgASgLMicuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxUZWNobmlxdWVIAIgBAUIUChJfY291bnNlbF90ZWNobmlxdWUi1QIKHVVwZGF0ZUNvdW5zZWxUZWNobmlxdWVSZXF1ZXN0EhwKFGNvdW5zZWxfdGVjaG5pcXVlX2lkGAEgASgJEhEKBG5hbWUYAiABKAlIAIgBARIUCgd0b25lX2lkGAMgASgJSAGIAQESEAoIaGFzX3RvbmUYBCABKAgSFwoKY29udGV4dF9pZBgFIAEoCUgCiAEBEhsKDmluc3RydWN0aW9uX2lkGAYgASgJSAOIAQESUgoXY291bnNlbF90ZWNobmlxdWVfc3RhZ2UYByABKA4yLC5jb20uaGVhcmxlcnMudjEubW9kZWwuQ291bnNlbFRlY2huaXF1ZVN0YWdlSASIAQFCBwoFX25hbWVCCgoIX3RvbmVfaWRCDQoLX2NvbnRleHRfaWRCEQoPX2luc3RydWN0aW9uX2lkQhoKGF9jb3Vuc2VsX3RlY2huaXF1ZV9zdGFnZSJkCh5VcGRhdGVDb3Vuc2VsVGVjaG5pcXVlUmVzcG9uc2USQgoRY291bnNlbF90ZWNobmlxdWUYASABKAsyJy5jb20uaGVhcmxlcnMudjEubW9kZWwuQ291bnNlbFRlY2huaXF1ZSJECiNTYXZlQ291bnNlbFRlY2huaXF1ZVNlcXVlbmNlUmVxdWVzdBIdChVjb3Vuc2VsX3RlY2huaXF1ZV9pZHMYASADKAkiawokU2F2ZUNvdW5zZWxUZWNobmlxdWVTZXF1ZW5jZVJlc3BvbnNlEkMKEmNvdW5zZWxfdGVjaG5pcXVlcxgBIAMoCzInLmNvbS5oZWFybGVycy52MS5tb2RlbC5Db3Vuc2VsVGVjaG5pcXVlIjoKFENyZWF0ZVBlcnNvbmFSZXF1ZXN0EhQKDGNvdW5zZWxvcl9pZBgBIAEoCRIMCgRib2R5GAIgASgJIkgKFUNyZWF0ZVBlcnNvbmFSZXNwb25zZRIvCgdwZXJzb25hGAEgASgLMh4uY29tLmhlYXJsZXJzLnYxLm1vZGVsLlBlcnNvbmEiQQoTRmluZFBlcnNvbmFzUmVxdWVzdBIZCgxjb3Vuc2Vsb3JfaWQYASABKAlIAIgBAUIPCg1fY291bnNlbG9yX2lkIkgKFEZpbmRQZXJzb25hc1Jlc3BvbnNlEjAKCHBlcnNvbmFzGAEgAygLMh4uY29tLmhlYXJsZXJzLnYxLm1vZGVsLlBlcnNvbmEiLAoWRmluZFBlcnNvbmFCeUlkUmVxdWVzdBISCgpwZXJzb25hX2lkGAEgASgJIlsKF0ZpbmRQZXJzb25hQnlJZFJlc3BvbnNlEjQKB3BlcnNvbmEYASABKAsyHi5jb20uaGVhcmxlcnMudjEubW9kZWwuUGVyc29uYUgAiAEBQgoKCF9wZXJzb25hIkYKFFVwZGF0ZVBlcnNvbmFSZXF1ZXN0EhIKCnBlcnNvbmFfaWQYASABKAkSEQoEYm9keRgCIAEoCUgAiAEBQgcKBV9ib2R5IkgKFVVwZGF0ZVBlcnNvbmFSZXNwb25zZRIvCgdwZXJzb25hGAEgASgLMh4uY29tLmhlYXJsZXJzLnYxLm1vZGVsLlBlcnNvbmEiSAoUQ3JlYXRlQ29udGV4dFJlcXVlc3QSDAoEbmFtZRgBIAEoCRIMCgRib2R5GAIgASgJEhQKDHBsYWNlaG9sZGVycxgDIAMoCSJIChVDcmVhdGVDb250ZXh0UmVzcG9uc2USLwoHY29udGV4dBgBIAEoCzIeLmNvbS5oZWFybGVycy52MS5tb2RlbC5Db250ZXh0IjEKE0ZpbmRDb250ZXh0c1JlcXVlc3QSEQoEbmFtZRgBIAEoCUgAiAEBQgcKBV9uYW1lIkgKFEZpbmRDb250ZXh0c1Jlc3BvbnNlEjAKCGNvbnRleHRzGAEgAygLMh4uY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvbnRleHQiLAoWRmluZENvbnRleHRCeUlkUmVxdWVzdBISCgpjb250ZXh0X2lkGAEgASgJIlsKF0ZpbmRDb250ZXh0QnlJZFJlc3BvbnNlEjQKB2NvbnRleHQYASABKAsyHi5jb20uaGVhcmxlcnMudjEubW9kZWwuQ29udGV4dEgAiAEBQgoKCF9jb250ZXh0IpIBChRVcGRhdGVDb250ZXh0UmVxdWVzdBISCgpjb250ZXh0X2lkGAEgASgJEhEKBG5hbWUYAiABKAlIAIgBARIRCgRib2R5GAMgASgJSAGIAQESFAoMcGxhY2Vob2xkZXJzGAQgAygJEhgKEGhhc19wbGFjZWhvbGRlcnMYBSABKAhCBwoFX25hbWVCBwoFX2JvZHkiSAoVVXBkYXRlQ29udGV4dFJlc3BvbnNlEi8KB2NvbnRleHQYASABKAsyHi5jb20uaGVhcmxlcnMudjEubW9kZWwuQ29udGV4dCJ6ChhDcmVhdGVJbnN0cnVjdGlvblJlcXVlc3QSDAoEbmFtZRgBIAEoCRIdChBpbml0aWFsX3NlbnRlbmNlGAIgASgJSACIAQESHAoUaW5zdHJ1Y3Rpb25faXRlbV9pZHMYAyADKAlCEwoRX2luaXRpYWxfc2VudGVuY2UiVAoZQ3JlYXRlSW5zdHJ1Y3Rpb25SZXNwb25zZRI3CgtpbnN0cnVjdGlvbhgBIAEoCzIiLmNvbS5oZWFybGVycy52MS5tb2RlbC5JbnN0cnVjdGlvbiI1ChdGaW5kSW5zdHJ1Y3Rpb25zUmVxdWVzdBIRCgRuYW1lGAEgASgJSACIAQFCBwoFX25hbWUiVAoYRmluZEluc3RydWN0aW9uc1Jlc3BvbnNlEjgKDGluc3RydWN0aW9ucxgBIAMoCzIiLmNvbS5oZWFybGVycy52MS5tb2RlbC5JbnN0cnVjdGlvbiI0ChpGaW5kSW5zdHJ1Y3Rpb25CeUlkUmVxdWVzdBIWCg5pbnN0cnVjdGlvbl9pZBgBIAEoCSJrChtGaW5kSW5zdHJ1Y3Rpb25CeUlkUmVzcG9uc2USPAoLaW5zdHJ1Y3Rpb24YASABKAsyIi5jb20uaGVhcmxlcnMudjEubW9kZWwuSW5zdHJ1Y3Rpb25IAIgBAUIOCgxfaW5zdHJ1Y3Rpb24i4AEKGFVwZGF0ZUluc3RydWN0aW9uUmVxdWVzdBIWCg5pbnN0cnVjdGlvbl9pZBgBIAEoCRIRCgRuYW1lGAIgASgJSACIAQESHQoQaW5pdGlhbF9zZW50ZW5jZRgDIAEoCUgBiAEBEhwKFGhhc19pbml0aWFsX3NlbnRlbmNlGAQgASgIEhwKFGluc3RydWN0aW9uX2l0ZW1faWRzGAUgAygJEiAKGGhhc19pbnN0cnVjdGlvbl9pdGVtX2lkcxgGIAEoCEIHCgVfbmFtZUITChFfaW5pdGlhbF9zZW50ZW5jZSJUChlVcGRhdGVJbnN0cnVjdGlvblJlc3BvbnNlEjcKC2luc3RydWN0aW9uGAEgASgLMiIuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkluc3RydWN0aW9uIiwKHENyZWF0ZUluc3RydWN0aW9uSXRlbVJlcXVlc3QSDAoEYm9keRgBIAEoCSJhCh1DcmVhdGVJbnN0cnVjdGlvbkl0ZW1SZXNwb25zZRJAChBpbnN0cnVjdGlvbl9pdGVtGAEgASgLMiYuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkluc3RydWN0aW9uSXRlbSI/ChtGaW5kSW5zdHJ1Y3Rpb25JdGVtc1JlcXVlc3QSFAoHa2V5d29yZBgBIAEoCUgAiAEBQgoKCF9rZXl3b3JkImEKHEZpbmRJbnN0cnVjdGlvbkl0ZW1zUmVzcG9uc2USQQoRaW5zdHJ1Y3Rpb25faXRlbXMYASADKAsyJi5jb20uaGVhcmxlcnMudjEubW9kZWwuSW5zdHJ1Y3Rpb25JdGVtIj0KHkZpbmRJbnN0cnVjdGlvbkl0ZW1CeUlkUmVxdWVzdBIbChNpbnN0cnVjdGlvbl9pdGVtX2lkGAEgASgJIn0KH0ZpbmRJbnN0cnVjdGlvbkl0ZW1CeUlkUmVzcG9uc2USRQoQaW5zdHJ1Y3Rpb25faXRlbRgBIAEoCzImLmNvbS5oZWFybGVycy52MS5tb2RlbC5JbnN0cnVjdGlvbkl0ZW1IAIgBAUITChFfaW5zdHJ1Y3Rpb25faXRlbSJXChxVcGRhdGVJbnN0cnVjdGlvbkl0ZW1SZXF1ZXN0EhsKE2luc3RydWN0aW9uX2l0ZW1faWQYASABKAkSEQoEYm9keRgCIAEoCUgAiAEBQgcKBV9ib2R5ImEKHVVwZGF0ZUluc3RydWN0aW9uSXRlbVJlc3BvbnNlEkAKEGluc3RydWN0aW9uX2l0ZW0YASABKAsyJi5jb20uaGVhcmxlcnMudjEubW9kZWwuSW5zdHJ1Y3Rpb25JdGVtIi8KEUNyZWF0ZVRvbmVSZXF1ZXN0EgwKBG5hbWUYASABKAkSDAoEYm9keRgCIAEoCSI/ChJDcmVhdGVUb25lUmVzcG9uc2USKQoEdG9uZRgBIAEoCzIbLmNvbS5oZWFybGVycy52MS5tb2RlbC5Ub25lIi4KEEZpbmRUb25lc1JlcXVlc3QSEQoEbmFtZRgBIAEoCUgAiAEBQgcKBV9uYW1lIj8KEUZpbmRUb25lc1Jlc3BvbnNlEioKBXRvbmVzGAEgAygLMhsuY29tLmhlYXJsZXJzLnYxLm1vZGVsLlRvbmUiJgoTRmluZFRvbmVCeUlkUmVxdWVzdBIPCgd0b25lX2lkGAEgASgJIk8KFEZpbmRUb25lQnlJZFJlc3BvbnNlEi4KBHRvbmUYASABKAsyGy5jb20uaGVhcmxlcnMudjEubW9kZWwuVG9uZUgAiAEBQgcKBV90b25lIlwKEVVwZGF0ZVRvbmVSZXF1ZXN0Eg8KB3RvbmVfaWQYASABKAkSEQoEbmFtZRgCIAEoCUgAiAEBEhEKBGJvZHkYAyABKAlIAYgBAUIHCgVfbmFtZUIHCgVfYm9keSI/ChJVcGRhdGVUb25lUmVzcG9uc2USKQoEdG9uZRgBIAEoCzIbLmNvbS5oZWFybGVycy52MS5tb2RlbC5Ub25lMr0iCg5Db3Vuc2VsU2VydmljZRJuCg1DcmVhdGVDb3Vuc2VsEi0uY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuQ3JlYXRlQ291bnNlbFJlcXVlc3QaLi5jb20uaGVhcmxlcnMudjEuc2VydmljZS5DcmVhdGVDb3Vuc2VsUmVzcG9uc2USawoMRmluZENvdW5zZWxzEiwuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZENvdW5zZWxzUmVxdWVzdBotLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkZpbmRDb3Vuc2Vsc1Jlc3BvbnNlEnQKD0ZpbmRDb3Vuc2VsQnlJZBIvLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkZpbmRDb3Vuc2VsQnlJZFJlcXVlc3QaMC5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kQ291bnNlbEJ5SWRSZXNwb25zZRJuCg1DcmVhdGVNZXNzYWdlEi0uY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuQ3JlYXRlTWVzc2FnZVJlcXVlc3QaLi5jb20uaGVhcmxlcnMudjEuc2VydmljZS5DcmVhdGVNZXNzYWdlUmVzcG9uc2USawoMRmluZE1lc3NhZ2VzEiwuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZE1lc3NhZ2VzUmVxdWVzdBotLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkZpbmRNZXNzYWdlc1Jlc3BvbnNlEmsKDFJlYWN0TWVzc2FnZRIsLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLlJlYWN0TWVzc2FnZVJlcXVlc3QaLS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5SZWFjdE1lc3NhZ2VSZXNwb25zZRJ0Cg9DcmVhdGVDb3Vuc2Vsb3ISLy5jb20uaGVhcmxlcnMudjEuc2VydmljZS5DcmVhdGVDb3Vuc2Vsb3JSZXF1ZXN0GjAuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuQ3JlYXRlQ291bnNlbG9yUmVzcG9uc2UScQoORmluZENvdW5zZWxvcnMSLi5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kQ291bnNlbG9yc1JlcXVlc3QaLy5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kQ291bnNlbG9yc1Jlc3BvbnNlEnoKEUZpbmRDb3Vuc2Vsb3JCeUlkEjEuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZENvdW5zZWxvckJ5SWRSZXF1ZXN0GjIuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZENvdW5zZWxvckJ5SWRSZXNwb25zZRJ0Cg9VcGRhdGVDb3Vuc2Vsb3ISLy5jb20uaGVhcmxlcnMudjEuc2VydmljZS5VcGRhdGVDb3Vuc2Vsb3JSZXF1ZXN0GjAuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuVXBkYXRlQ291bnNlbG9yUmVzcG9uc2USoQEKHkZpbmRDb3Vuc2Vsb3JVc2VyUmVsYXRpb25zaGlwcxI+LmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkZpbmRDb3Vuc2Vsb3JVc2VyUmVsYXRpb25zaGlwc1JlcXVlc3QaPy5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kQ291bnNlbG9yVXNlclJlbGF0aW9uc2hpcHNSZXNwb25zZRKJAQoWQ3JlYXRlQ291bnNlbFRlY2huaXF1ZRI2LmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkNyZWF0ZUNvdW5zZWxUZWNobmlxdWVSZXF1ZXN0GjcuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuQ3JlYXRlQ291bnNlbFRlY2huaXF1ZVJlc3BvbnNlEoYBChVGaW5kQ291bnNlbFRlY2huaXF1ZXMSNS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kQ291bnNlbFRlY2huaXF1ZXNSZXF1ZXN0GjYuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZENvdW5zZWxUZWNobmlxdWVzUmVzcG9uc2USjwEKGEZpbmRDb3Vuc2VsVGVjaG5pcXVlQnlJZBI4LmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkZpbmRDb3Vuc2VsVGVjaG5pcXVlQnlJZFJlcXVlc3QaOS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kQ291bnNlbFRlY2huaXF1ZUJ5SWRSZXNwb25zZRKJAQoWVXBkYXRlQ291bnNlbFRlY2huaXF1ZRI2LmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLlVwZGF0ZUNvdW5zZWxUZWNobmlxdWVSZXF1ZXN0GjcuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuVXBkYXRlQ291bnNlbFRlY2huaXF1ZVJlc3BvbnNlEpsBChxTYXZlQ291bnNlbFRlY2huaXF1ZVNlcXVlbmNlEjwuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuU2F2ZUNvdW5zZWxUZWNobmlxdWVTZXF1ZW5jZVJlcXVlc3QaPS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5TYXZlQ291bnNlbFRlY2huaXF1ZVNlcXVlbmNlUmVzcG9uc2USbgoNQ3JlYXRlUGVyc29uYRItLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkNyZWF0ZVBlcnNvbmFSZXF1ZXN0Gi4uY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuQ3JlYXRlUGVyc29uYVJlc3BvbnNlEmsKDEZpbmRQZXJzb25hcxIsLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkZpbmRQZXJzb25hc1JlcXVlc3QaLS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kUGVyc29uYXNSZXNwb25zZRJ0Cg9GaW5kUGVyc29uYUJ5SWQSLy5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kUGVyc29uYUJ5SWRSZXF1ZXN0GjAuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZFBlcnNvbmFCeUlkUmVzcG9uc2USbgoNVXBkYXRlUGVyc29uYRItLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLlVwZGF0ZVBlcnNvbmFSZXF1ZXN0Gi4uY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuVXBkYXRlUGVyc29uYVJlc3BvbnNlEm4KDUNyZWF0ZUNvbnRleHQSLS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5DcmVhdGVDb250ZXh0UmVxdWVzdBouLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkNyZWF0ZUNvbnRleHRSZXNwb25zZRJrCgxGaW5kQ29udGV4dHMSLC5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kQ29udGV4dHNSZXF1ZXN0Gi0uY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZENvbnRleHRzUmVzcG9uc2USdAoPRmluZENvbnRleHRCeUlkEi8uY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZENvbnRleHRCeUlkUmVxdWVzdBowLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkZpbmRDb250ZXh0QnlJZFJlc3BvbnNlEm4KDVVwZGF0ZUNvbnRleHQSLS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5VcGRhdGVDb250ZXh0UmVxdWVzdBouLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLlVwZGF0ZUNvbnRleHRSZXNwb25zZRJ6ChFDcmVhdGVJbnN0cnVjdGlvbhIxLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkNyZWF0ZUluc3RydWN0aW9uUmVxdWVzdBoyLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkNyZWF0ZUluc3RydWN0aW9uUmVzcG9uc2USdwoQRmluZEluc3RydWN0aW9ucxIwLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkZpbmRJbnN0cnVjdGlvbnNSZXF1ZXN0GjEuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZEluc3RydWN0aW9uc1Jlc3BvbnNlEoABChNGaW5kSW5zdHJ1Y3Rpb25CeUlkEjMuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZEluc3RydWN0aW9uQnlJZFJlcXVlc3QaNC5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kSW5zdHJ1Y3Rpb25CeUlkUmVzcG9uc2USegoRVXBkYXRlSW5zdHJ1Y3Rpb24SMS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5VcGRhdGVJbnN0cnVjdGlvblJlcXVlc3QaMi5jb20uaGVhcmxlcnMudjEuc2VydmljZS5VcGRhdGVJbnN0cnVjdGlvblJlc3BvbnNlEoYBChVDcmVhdGVJbnN0cnVjdGlvbkl0ZW0SNS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5DcmVhdGVJbnN0cnVjdGlvbkl0ZW1SZXF1ZXN0GjYuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuQ3JlYXRlSW5zdHJ1Y3Rpb25JdGVtUmVzcG9uc2USgwEKFEZpbmRJbnN0cnVjdGlvbkl0ZW1zEjQuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZEluc3RydWN0aW9uSXRlbXNSZXF1ZXN0GjUuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZEluc3RydWN0aW9uSXRlbXNSZXNwb25zZRKMAQoXRmluZEluc3RydWN0aW9uSXRlbUJ5SWQSNy5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kSW5zdHJ1Y3Rpb25JdGVtQnlJZFJlcXVlc3QaOC5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kSW5zdHJ1Y3Rpb25JdGVtQnlJZFJlc3BvbnNlEoYBChVVcGRhdGVJbnN0cnVjdGlvbkl0ZW0SNS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5VcGRhdGVJbnN0cnVjdGlvbkl0ZW1SZXF1ZXN0GjYuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuVXBkYXRlSW5zdHJ1Y3Rpb25JdGVtUmVzcG9uc2USZQoKQ3JlYXRlVG9uZRIqLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkNyZWF0ZVRvbmVSZXF1ZXN0GisuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuQ3JlYXRlVG9uZVJlc3BvbnNlEmIKCUZpbmRUb25lcxIpLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkZpbmRUb25lc1JlcXVlc3QaKi5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kVG9uZXNSZXNwb25zZRJrCgxGaW5kVG9uZUJ5SWQSLC5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kVG9uZUJ5SWRSZXF1ZXN0Gi0uY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZFRvbmVCeUlkUmVzcG9uc2USZQoKVXBkYXRlVG9uZRIqLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLlVwZGF0ZVRvbmVSZXF1ZXN0GisuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuVXBkYXRlVG9uZVJlc3BvbnNlQqsBChtjb20uY29tLmhlYXJsZXJzLnYxLnNlcnZpY2VCDENvdW5zZWxQcm90b1ABogIEQ0hWU6oCF0NvbS5IZWFybGVycy5WMS5TZXJ2aWNlygIXQ29tXEhlYXJsZXJzXFYxXFNlcnZpY2XiAiNDb21cSGVhcmxlcnNcVjFcU2VydmljZVxHUEJNZXRhZGF0YeoCGkNvbTo6SGVhcmxlcnM6OlYxOjpTZXJ2aWNlYgZwcm90bzM", [file_com_hearlers_v1_model_counsel]);
+  fileDesc("CiVjb20vaGVhcmxlcnMvdjEvc2VydmljZS9jb3Vuc2VsLnByb3RvEhdjb20uaGVhcmxlcnMudjEuc2VydmljZSKfAQoUQ3JlYXRlQ291bnNlbFJlcXVlc3QSDwoHdXNlcl9pZBgBIAEoCRIUCgxjb3Vuc2Vsb3JfaWQYAiABKAkSGgoNaW50cm9fbWVzc2FnZRgDIAEoCUgAiAEBEh0KEHJlc3BvbnNlX21lc3NhZ2UYBCABKAlIAYgBAUIQCg5faW50cm9fbWVzc2FnZUITChFfcmVzcG9uc2VfbWVzc2FnZSKJAQoVQ3JlYXRlQ291bnNlbFJlc3BvbnNlEi8KB2NvdW5zZWwYASABKAsyHi5jb20uaGVhcmxlcnMudjEubW9kZWwuQ291bnNlbBI/ChBjb3Vuc2VsX21lc3NhZ2VzGAIgAygLMiUuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxNZXNzYWdlIlIKE0ZpbmRDb3Vuc2Vsc1JlcXVlc3QSDwoHdXNlcl9pZBgBIAEoCRIZCgxjb3Vuc2Vsb3JfaWQYAiABKAlIAIgBAUIPCg1fY291bnNlbG9yX2lkIkgKFEZpbmRDb3Vuc2Vsc1Jlc3BvbnNlEjAKCGNvdW5zZWxzGAEgAygLMh4uY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWwiLAoWRmluZENvdW5zZWxCeUlkUmVxdWVzdBISCgpjb3Vuc2VsX2lkGAEgASgJIlsKF0ZpbmRDb3Vuc2VsQnlJZFJlc3BvbnNlEjQKB2NvdW5zZWwYASABKAsyHi5jb20uaGVhcmxlcnMudjEubW9kZWwuQ291bnNlbEgAiAEBQgoKCF9jb3Vuc2VsIjsKFENyZWF0ZU1lc3NhZ2VSZXF1ZXN0EhIKCmNvdW5zZWxfaWQYASABKAkSDwoHbWVzc2FnZRgCIAEoCSKqAQoVQ3JlYXRlTWVzc2FnZVJlc3BvbnNlEkYKF2NyZWF0ZWRfY291bnNlbF9tZXNzYWdlGAEgASgLMiUuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxNZXNzYWdlEkkKGmNvdW5zZWxvcl9yZXNwb25zZV9tZXNzYWdlGAIgASgLMiUuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxNZXNzYWdlIikKE0ZpbmRNZXNzYWdlc1JlcXVlc3QSEgoKY291bnNlbF9pZBgBIAEoCSJXChRGaW5kTWVzc2FnZXNSZXNwb25zZRI/ChBjb3Vuc2VsX21lc3NhZ2VzGAEgAygLMiUuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxNZXNzYWdlImoKE1JlYWN0TWVzc2FnZVJlcXVlc3QSEgoKbWVzc2FnZV9pZBgBIAEoCRI/CghyZWFjdGlvbhgCIAEoDjItLmNvbS5oZWFybGVycy52MS5tb2RlbC5Db3Vuc2VsTWVzc2FnZVJlYWN0aW9uIlYKFFJlYWN0TWVzc2FnZVJlc3BvbnNlEj4KD2NvdW5zZWxfbWVzc2FnZRgBIAEoCzIlLmNvbS5oZWFybGVycy52MS5tb2RlbC5Db3Vuc2VsTWVzc2FnZSJOCiVGaW5kQ291bnNlbG9yVXNlclJlbGF0aW9uc2hpcHNSZXF1ZXN0EhQKDGNvdW5zZWxvcl9pZBgBIAEoCRIPCgd1c2VyX2lkGAIgASgJIoABCiZGaW5kQ291bnNlbG9yVXNlclJlbGF0aW9uc2hpcHNSZXNwb25zZRJWChxjb3Vuc2Vsb3JfdXNlcl9yZWxhdGlvbnNoaXBzGAEgAygLMjAuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxvclVzZXJSZWxhdGlvbnNoaXAy0QYKDkNvdW5zZWxTZXJ2aWNlEm4KDUNyZWF0ZUNvdW5zZWwSLS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5DcmVhdGVDb3Vuc2VsUmVxdWVzdBouLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkNyZWF0ZUNvdW5zZWxSZXNwb25zZRJrCgxGaW5kQ291bnNlbHMSLC5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kQ291bnNlbHNSZXF1ZXN0Gi0uY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZENvdW5zZWxzUmVzcG9uc2USdAoPRmluZENvdW5zZWxCeUlkEi8uY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZENvdW5zZWxCeUlkUmVxdWVzdBowLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkZpbmRDb3Vuc2VsQnlJZFJlc3BvbnNlEm4KDUNyZWF0ZU1lc3NhZ2USLS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5DcmVhdGVNZXNzYWdlUmVxdWVzdBouLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkNyZWF0ZU1lc3NhZ2VSZXNwb25zZRJrCgxGaW5kTWVzc2FnZXMSLC5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kTWVzc2FnZXNSZXF1ZXN0Gi0uY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZE1lc3NhZ2VzUmVzcG9uc2USawoMUmVhY3RNZXNzYWdlEiwuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuUmVhY3RNZXNzYWdlUmVxdWVzdBotLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLlJlYWN0TWVzc2FnZVJlc3BvbnNlEqEBCh5GaW5kQ291bnNlbG9yVXNlclJlbGF0aW9uc2hpcHMSPi5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kQ291bnNlbG9yVXNlclJlbGF0aW9uc2hpcHNSZXF1ZXN0Gj8uY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZENvdW5zZWxvclVzZXJSZWxhdGlvbnNoaXBzUmVzcG9uc2VCqwEKG2NvbS5jb20uaGVhcmxlcnMudjEuc2VydmljZUIMQ291bnNlbFByb3RvUAGiAgRDSFZTqgIXQ29tLkhlYXJsZXJzLlYxLlNlcnZpY2XKAhdDb21cSGVhcmxlcnNcVjFcU2VydmljZeICI0NvbVxIZWFybGVyc1xWMVxTZXJ2aWNlXEdQQk1ldGFkYXRh6gIaQ29tOjpIZWFybGVyczo6VjE6OlNlcnZpY2ViBnByb3RvMw", [file_com_hearlers_v1_model_counsel]);
 
 /**
  * 상담
@@ -263,179 +263,6 @@ export const ReactMessageResponseSchema: GenMessage<ReactMessageResponse> = /*@_
   messageDesc(file_com_hearlers_v1_service_counsel, 11);
 
 /**
- * 상담사
- *
- * @generated from message com.hearlers.v1.service.CreateCounselorRequest
- */
-export type CreateCounselorRequest = Message<"com.hearlers.v1.service.CreateCounselorRequest"> & {
-  /**
-   * @generated from field: string tone_id = 1;
-   */
-  toneId: string;
-
-  /**
-   * @generated from field: string name = 2;
-   */
-  name: string;
-
-  /**
-   * @generated from field: string description = 3;
-   */
-  description: string;
-
-  /**
-   * @generated from field: com.hearlers.v1.model.CounselorGender counselor_gender = 4;
-   */
-  counselorGender: CounselorGender;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.CreateCounselorRequest.
- * Use `create(CreateCounselorRequestSchema)` to create a new message.
- */
-export const CreateCounselorRequestSchema: GenMessage<CreateCounselorRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 12);
-
-/**
- * @generated from message com.hearlers.v1.service.CreateCounselorResponse
- */
-export type CreateCounselorResponse = Message<"com.hearlers.v1.service.CreateCounselorResponse"> & {
-  /**
-   * @generated from field: com.hearlers.v1.model.Counselor counselor = 1;
-   */
-  counselor?: Counselor;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.CreateCounselorResponse.
- * Use `create(CreateCounselorResponseSchema)` to create a new message.
- */
-export const CreateCounselorResponseSchema: GenMessage<CreateCounselorResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 13);
-
-/**
- * @generated from message com.hearlers.v1.service.FindCounselorsRequest
- */
-export type FindCounselorsRequest = Message<"com.hearlers.v1.service.FindCounselorsRequest"> & {
-  /**
-   * @generated from field: optional string tone_id = 1;
-   */
-  toneId?: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindCounselorsRequest.
- * Use `create(FindCounselorsRequestSchema)` to create a new message.
- */
-export const FindCounselorsRequestSchema: GenMessage<FindCounselorsRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 14);
-
-/**
- * @generated from message com.hearlers.v1.service.FindCounselorsResponse
- */
-export type FindCounselorsResponse = Message<"com.hearlers.v1.service.FindCounselorsResponse"> & {
-  /**
-   * @generated from field: repeated com.hearlers.v1.model.Counselor counselors = 1;
-   */
-  counselors: Counselor[];
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindCounselorsResponse.
- * Use `create(FindCounselorsResponseSchema)` to create a new message.
- */
-export const FindCounselorsResponseSchema: GenMessage<FindCounselorsResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 15);
-
-/**
- * @generated from message com.hearlers.v1.service.FindCounselorByIdRequest
- */
-export type FindCounselorByIdRequest = Message<"com.hearlers.v1.service.FindCounselorByIdRequest"> & {
-  /**
-   * @generated from field: string counselor_id = 1;
-   */
-  counselorId: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindCounselorByIdRequest.
- * Use `create(FindCounselorByIdRequestSchema)` to create a new message.
- */
-export const FindCounselorByIdRequestSchema: GenMessage<FindCounselorByIdRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 16);
-
-/**
- * @generated from message com.hearlers.v1.service.FindCounselorByIdResponse
- */
-export type FindCounselorByIdResponse = Message<"com.hearlers.v1.service.FindCounselorByIdResponse"> & {
-  /**
-   * @generated from field: optional com.hearlers.v1.model.Counselor counselor = 1;
-   */
-  counselor?: Counselor;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindCounselorByIdResponse.
- * Use `create(FindCounselorByIdResponseSchema)` to create a new message.
- */
-export const FindCounselorByIdResponseSchema: GenMessage<FindCounselorByIdResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 17);
-
-/**
- * @generated from message com.hearlers.v1.service.UpdateCounselorRequest
- */
-export type UpdateCounselorRequest = Message<"com.hearlers.v1.service.UpdateCounselorRequest"> & {
-  /**
-   * @generated from field: string counselor_id = 1;
-   */
-  counselorId: string;
-
-  /**
-   * @generated from field: optional string tone_id = 2;
-   */
-  toneId?: string;
-
-  /**
-   * @generated from field: optional string name = 3;
-   */
-  name?: string;
-
-  /**
-   * @generated from field: optional string description = 4;
-   */
-  description?: string;
-
-  /**
-   * @generated from field: optional com.hearlers.v1.model.CounselorGender counselor_gender = 5;
-   */
-  counselorGender?: CounselorGender;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.UpdateCounselorRequest.
- * Use `create(UpdateCounselorRequestSchema)` to create a new message.
- */
-export const UpdateCounselorRequestSchema: GenMessage<UpdateCounselorRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 18);
-
-/**
- * @generated from message com.hearlers.v1.service.UpdateCounselorResponse
- */
-export type UpdateCounselorResponse = Message<"com.hearlers.v1.service.UpdateCounselorResponse"> & {
-  /**
-   * @generated from field: com.hearlers.v1.model.Counselor counselor = 1;
-   */
-  counselor?: Counselor;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.UpdateCounselorResponse.
- * Use `create(UpdateCounselorResponseSchema)` to create a new message.
- */
-export const UpdateCounselorResponseSchema: GenMessage<UpdateCounselorResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 19);
-
-/**
  * 상담사와 유저의 관계
  *
  * @generated from message com.hearlers.v1.service.FindCounselorUserRelationshipsRequest
@@ -457,7 +284,7 @@ export type FindCounselorUserRelationshipsRequest = Message<"com.hearlers.v1.ser
  * Use `create(FindCounselorUserRelationshipsRequestSchema)` to create a new message.
  */
 export const FindCounselorUserRelationshipsRequestSchema: GenMessage<FindCounselorUserRelationshipsRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 20);
+  messageDesc(file_com_hearlers_v1_service_counsel, 12);
 
 /**
  * @generated from message com.hearlers.v1.service.FindCounselorUserRelationshipsResponse
@@ -474,1034 +301,7 @@ export type FindCounselorUserRelationshipsResponse = Message<"com.hearlers.v1.se
  * Use `create(FindCounselorUserRelationshipsResponseSchema)` to create a new message.
  */
 export const FindCounselorUserRelationshipsResponseSchema: GenMessage<FindCounselorUserRelationshipsResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 21);
-
-/**
- * 상담 기법
- *
- * @generated from message com.hearlers.v1.service.CreateCounselTechniqueRequest
- */
-export type CreateCounselTechniqueRequest = Message<"com.hearlers.v1.service.CreateCounselTechniqueRequest"> & {
-  /**
-   * @generated from field: string name = 1;
-   */
-  name: string;
-
-  /**
-   * @generated from field: optional string tone_id = 2;
-   */
-  toneId?: string;
-
-  /**
-   * @generated from field: string context_id = 3;
-   */
-  contextId: string;
-
-  /**
-   * @generated from field: string instruction_id = 4;
-   */
-  instructionId: string;
-
-  /**
-   * @generated from field: com.hearlers.v1.model.CounselTechniqueStage counsel_technique_stage = 5;
-   */
-  counselTechniqueStage: CounselTechniqueStage;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.CreateCounselTechniqueRequest.
- * Use `create(CreateCounselTechniqueRequestSchema)` to create a new message.
- */
-export const CreateCounselTechniqueRequestSchema: GenMessage<CreateCounselTechniqueRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 22);
-
-/**
- * @generated from message com.hearlers.v1.service.CreateCounselTechniqueResponse
- */
-export type CreateCounselTechniqueResponse = Message<"com.hearlers.v1.service.CreateCounselTechniqueResponse"> & {
-  /**
-   * @generated from field: com.hearlers.v1.model.CounselTechnique counsel_technique = 1;
-   */
-  counselTechnique?: CounselTechnique;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.CreateCounselTechniqueResponse.
- * Use `create(CreateCounselTechniqueResponseSchema)` to create a new message.
- */
-export const CreateCounselTechniqueResponseSchema: GenMessage<CreateCounselTechniqueResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 23);
-
-/**
- * @generated from message com.hearlers.v1.service.FindCounselTechniquesRequest
- */
-export type FindCounselTechniquesRequest = Message<"com.hearlers.v1.service.FindCounselTechniquesRequest"> & {
-  /**
-   * @generated from field: optional string name = 1;
-   */
-  name?: string;
-
-  /**
-   * @generated from field: optional string tone_id = 2;
-   */
-  toneId?: string;
-
-  /**
-   * @generated from field: optional com.hearlers.v1.model.CounselTechniqueStage counsel_technique_stage = 3;
-   */
-  counselTechniqueStage?: CounselTechniqueStage;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindCounselTechniquesRequest.
- * Use `create(FindCounselTechniquesRequestSchema)` to create a new message.
- */
-export const FindCounselTechniquesRequestSchema: GenMessage<FindCounselTechniquesRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 24);
-
-/**
- * @generated from message com.hearlers.v1.service.FindCounselTechniquesResponse
- */
-export type FindCounselTechniquesResponse = Message<"com.hearlers.v1.service.FindCounselTechniquesResponse"> & {
-  /**
-   * @generated from field: repeated com.hearlers.v1.model.CounselTechnique counsel_techniques = 1;
-   */
-  counselTechniques: CounselTechnique[];
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindCounselTechniquesResponse.
- * Use `create(FindCounselTechniquesResponseSchema)` to create a new message.
- */
-export const FindCounselTechniquesResponseSchema: GenMessage<FindCounselTechniquesResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 25);
-
-/**
- * @generated from message com.hearlers.v1.service.FindCounselTechniqueByIdRequest
- */
-export type FindCounselTechniqueByIdRequest = Message<"com.hearlers.v1.service.FindCounselTechniqueByIdRequest"> & {
-  /**
-   * @generated from field: string counsel_technique_id = 1;
-   */
-  counselTechniqueId: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindCounselTechniqueByIdRequest.
- * Use `create(FindCounselTechniqueByIdRequestSchema)` to create a new message.
- */
-export const FindCounselTechniqueByIdRequestSchema: GenMessage<FindCounselTechniqueByIdRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 26);
-
-/**
- * @generated from message com.hearlers.v1.service.FindCounselTechniqueByIdResponse
- */
-export type FindCounselTechniqueByIdResponse = Message<"com.hearlers.v1.service.FindCounselTechniqueByIdResponse"> & {
-  /**
-   * @generated from field: optional com.hearlers.v1.model.CounselTechnique counsel_technique = 1;
-   */
-  counselTechnique?: CounselTechnique;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindCounselTechniqueByIdResponse.
- * Use `create(FindCounselTechniqueByIdResponseSchema)` to create a new message.
- */
-export const FindCounselTechniqueByIdResponseSchema: GenMessage<FindCounselTechniqueByIdResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 27);
-
-/**
- * Tone Id 의 경우 nullable field
- * Tone Id 의 변경이 없을 경우 has_tone = false
- * Tone Id 를 null 로 변경할 경우 has_tone = true, tone_id = null
- * Tone Id 를 다른 값으로 변경할 경우 has_tone = true, tone_id = "new_tone_id"
- *
- * @generated from message com.hearlers.v1.service.UpdateCounselTechniqueRequest
- */
-export type UpdateCounselTechniqueRequest = Message<"com.hearlers.v1.service.UpdateCounselTechniqueRequest"> & {
-  /**
-   * @generated from field: string counsel_technique_id = 1;
-   */
-  counselTechniqueId: string;
-
-  /**
-   * @generated from field: optional string name = 2;
-   */
-  name?: string;
-
-  /**
-   * @generated from field: optional string tone_id = 3;
-   */
-  toneId?: string;
-
-  /**
-   * @generated from field: bool has_tone = 4;
-   */
-  hasTone: boolean;
-
-  /**
-   * @generated from field: optional string context_id = 5;
-   */
-  contextId?: string;
-
-  /**
-   * @generated from field: optional string instruction_id = 6;
-   */
-  instructionId?: string;
-
-  /**
-   * @generated from field: optional com.hearlers.v1.model.CounselTechniqueStage counsel_technique_stage = 7;
-   */
-  counselTechniqueStage?: CounselTechniqueStage;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.UpdateCounselTechniqueRequest.
- * Use `create(UpdateCounselTechniqueRequestSchema)` to create a new message.
- */
-export const UpdateCounselTechniqueRequestSchema: GenMessage<UpdateCounselTechniqueRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 28);
-
-/**
- * @generated from message com.hearlers.v1.service.UpdateCounselTechniqueResponse
- */
-export type UpdateCounselTechniqueResponse = Message<"com.hearlers.v1.service.UpdateCounselTechniqueResponse"> & {
-  /**
-   * @generated from field: com.hearlers.v1.model.CounselTechnique counsel_technique = 1;
-   */
-  counselTechnique?: CounselTechnique;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.UpdateCounselTechniqueResponse.
- * Use `create(UpdateCounselTechniqueResponseSchema)` to create a new message.
- */
-export const UpdateCounselTechniqueResponseSchema: GenMessage<UpdateCounselTechniqueResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 29);
-
-/**
- * @generated from message com.hearlers.v1.service.SaveCounselTechniqueSequenceRequest
- */
-export type SaveCounselTechniqueSequenceRequest = Message<"com.hearlers.v1.service.SaveCounselTechniqueSequenceRequest"> & {
-  /**
-   * @generated from field: repeated string counsel_technique_ids = 1;
-   */
-  counselTechniqueIds: string[];
-};
-
-/**
- * Describes the message com.hearlers.v1.service.SaveCounselTechniqueSequenceRequest.
- * Use `create(SaveCounselTechniqueSequenceRequestSchema)` to create a new message.
- */
-export const SaveCounselTechniqueSequenceRequestSchema: GenMessage<SaveCounselTechniqueSequenceRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 30);
-
-/**
- * @generated from message com.hearlers.v1.service.SaveCounselTechniqueSequenceResponse
- */
-export type SaveCounselTechniqueSequenceResponse = Message<"com.hearlers.v1.service.SaveCounselTechniqueSequenceResponse"> & {
-  /**
-   * @generated from field: repeated com.hearlers.v1.model.CounselTechnique counsel_techniques = 1;
-   */
-  counselTechniques: CounselTechnique[];
-};
-
-/**
- * Describes the message com.hearlers.v1.service.SaveCounselTechniqueSequenceResponse.
- * Use `create(SaveCounselTechniqueSequenceResponseSchema)` to create a new message.
- */
-export const SaveCounselTechniqueSequenceResponseSchema: GenMessage<SaveCounselTechniqueSequenceResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 31);
-
-/**
- * Persona
- *
- * @generated from message com.hearlers.v1.service.CreatePersonaRequest
- */
-export type CreatePersonaRequest = Message<"com.hearlers.v1.service.CreatePersonaRequest"> & {
-  /**
-   * @generated from field: string counselor_id = 1;
-   */
-  counselorId: string;
-
-  /**
-   * @generated from field: string body = 2;
-   */
-  body: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.CreatePersonaRequest.
- * Use `create(CreatePersonaRequestSchema)` to create a new message.
- */
-export const CreatePersonaRequestSchema: GenMessage<CreatePersonaRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 32);
-
-/**
- * @generated from message com.hearlers.v1.service.CreatePersonaResponse
- */
-export type CreatePersonaResponse = Message<"com.hearlers.v1.service.CreatePersonaResponse"> & {
-  /**
-   * @generated from field: com.hearlers.v1.model.Persona persona = 1;
-   */
-  persona?: Persona;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.CreatePersonaResponse.
- * Use `create(CreatePersonaResponseSchema)` to create a new message.
- */
-export const CreatePersonaResponseSchema: GenMessage<CreatePersonaResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 33);
-
-/**
- * @generated from message com.hearlers.v1.service.FindPersonasRequest
- */
-export type FindPersonasRequest = Message<"com.hearlers.v1.service.FindPersonasRequest"> & {
-  /**
-   * @generated from field: optional string counselor_id = 1;
-   */
-  counselorId?: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindPersonasRequest.
- * Use `create(FindPersonasRequestSchema)` to create a new message.
- */
-export const FindPersonasRequestSchema: GenMessage<FindPersonasRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 34);
-
-/**
- * @generated from message com.hearlers.v1.service.FindPersonasResponse
- */
-export type FindPersonasResponse = Message<"com.hearlers.v1.service.FindPersonasResponse"> & {
-  /**
-   * @generated from field: repeated com.hearlers.v1.model.Persona personas = 1;
-   */
-  personas: Persona[];
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindPersonasResponse.
- * Use `create(FindPersonasResponseSchema)` to create a new message.
- */
-export const FindPersonasResponseSchema: GenMessage<FindPersonasResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 35);
-
-/**
- * @generated from message com.hearlers.v1.service.FindPersonaByIdRequest
- */
-export type FindPersonaByIdRequest = Message<"com.hearlers.v1.service.FindPersonaByIdRequest"> & {
-  /**
-   * @generated from field: string persona_id = 1;
-   */
-  personaId: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindPersonaByIdRequest.
- * Use `create(FindPersonaByIdRequestSchema)` to create a new message.
- */
-export const FindPersonaByIdRequestSchema: GenMessage<FindPersonaByIdRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 36);
-
-/**
- * @generated from message com.hearlers.v1.service.FindPersonaByIdResponse
- */
-export type FindPersonaByIdResponse = Message<"com.hearlers.v1.service.FindPersonaByIdResponse"> & {
-  /**
-   * @generated from field: optional com.hearlers.v1.model.Persona persona = 1;
-   */
-  persona?: Persona;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindPersonaByIdResponse.
- * Use `create(FindPersonaByIdResponseSchema)` to create a new message.
- */
-export const FindPersonaByIdResponseSchema: GenMessage<FindPersonaByIdResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 37);
-
-/**
- * @generated from message com.hearlers.v1.service.UpdatePersonaRequest
- */
-export type UpdatePersonaRequest = Message<"com.hearlers.v1.service.UpdatePersonaRequest"> & {
-  /**
-   * @generated from field: string persona_id = 1;
-   */
-  personaId: string;
-
-  /**
-   * @generated from field: optional string body = 2;
-   */
-  body?: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.UpdatePersonaRequest.
- * Use `create(UpdatePersonaRequestSchema)` to create a new message.
- */
-export const UpdatePersonaRequestSchema: GenMessage<UpdatePersonaRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 38);
-
-/**
- * @generated from message com.hearlers.v1.service.UpdatePersonaResponse
- */
-export type UpdatePersonaResponse = Message<"com.hearlers.v1.service.UpdatePersonaResponse"> & {
-  /**
-   * @generated from field: com.hearlers.v1.model.Persona persona = 1;
-   */
-  persona?: Persona;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.UpdatePersonaResponse.
- * Use `create(UpdatePersonaResponseSchema)` to create a new message.
- */
-export const UpdatePersonaResponseSchema: GenMessage<UpdatePersonaResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 39);
-
-/**
- * Context
- *
- * @generated from message com.hearlers.v1.service.CreateContextRequest
- */
-export type CreateContextRequest = Message<"com.hearlers.v1.service.CreateContextRequest"> & {
-  /**
-   * @generated from field: string name = 1;
-   */
-  name: string;
-
-  /**
-   * @generated from field: string body = 2;
-   */
-  body: string;
-
-  /**
-   * @generated from field: repeated string placeholders = 3;
-   */
-  placeholders: string[];
-};
-
-/**
- * Describes the message com.hearlers.v1.service.CreateContextRequest.
- * Use `create(CreateContextRequestSchema)` to create a new message.
- */
-export const CreateContextRequestSchema: GenMessage<CreateContextRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 40);
-
-/**
- * @generated from message com.hearlers.v1.service.CreateContextResponse
- */
-export type CreateContextResponse = Message<"com.hearlers.v1.service.CreateContextResponse"> & {
-  /**
-   * @generated from field: com.hearlers.v1.model.Context context = 1;
-   */
-  context?: Context;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.CreateContextResponse.
- * Use `create(CreateContextResponseSchema)` to create a new message.
- */
-export const CreateContextResponseSchema: GenMessage<CreateContextResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 41);
-
-/**
- * @generated from message com.hearlers.v1.service.FindContextsRequest
- */
-export type FindContextsRequest = Message<"com.hearlers.v1.service.FindContextsRequest"> & {
-  /**
-   * @generated from field: optional string name = 1;
-   */
-  name?: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindContextsRequest.
- * Use `create(FindContextsRequestSchema)` to create a new message.
- */
-export const FindContextsRequestSchema: GenMessage<FindContextsRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 42);
-
-/**
- * @generated from message com.hearlers.v1.service.FindContextsResponse
- */
-export type FindContextsResponse = Message<"com.hearlers.v1.service.FindContextsResponse"> & {
-  /**
-   * @generated from field: repeated com.hearlers.v1.model.Context contexts = 1;
-   */
-  contexts: Context[];
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindContextsResponse.
- * Use `create(FindContextsResponseSchema)` to create a new message.
- */
-export const FindContextsResponseSchema: GenMessage<FindContextsResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 43);
-
-/**
- * @generated from message com.hearlers.v1.service.FindContextByIdRequest
- */
-export type FindContextByIdRequest = Message<"com.hearlers.v1.service.FindContextByIdRequest"> & {
-  /**
-   * @generated from field: string context_id = 1;
-   */
-  contextId: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindContextByIdRequest.
- * Use `create(FindContextByIdRequestSchema)` to create a new message.
- */
-export const FindContextByIdRequestSchema: GenMessage<FindContextByIdRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 44);
-
-/**
- * @generated from message com.hearlers.v1.service.FindContextByIdResponse
- */
-export type FindContextByIdResponse = Message<"com.hearlers.v1.service.FindContextByIdResponse"> & {
-  /**
-   * @generated from field: optional com.hearlers.v1.model.Context context = 1;
-   */
-  context?: Context;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindContextByIdResponse.
- * Use `create(FindContextByIdResponseSchema)` to create a new message.
- */
-export const FindContextByIdResponseSchema: GenMessage<FindContextByIdResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 45);
-
-/**
- * @generated from message com.hearlers.v1.service.UpdateContextRequest
- */
-export type UpdateContextRequest = Message<"com.hearlers.v1.service.UpdateContextRequest"> & {
-  /**
-   * @generated from field: string context_id = 1;
-   */
-  contextId: string;
-
-  /**
-   * @generated from field: optional string name = 2;
-   */
-  name?: string;
-
-  /**
-   * @generated from field: optional string body = 3;
-   */
-  body?: string;
-
-  /**
-   * @generated from field: repeated string placeholders = 4;
-   */
-  placeholders: string[];
-
-  /**
-   * @generated from field: bool has_placeholders = 5;
-   */
-  hasPlaceholders: boolean;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.UpdateContextRequest.
- * Use `create(UpdateContextRequestSchema)` to create a new message.
- */
-export const UpdateContextRequestSchema: GenMessage<UpdateContextRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 46);
-
-/**
- * @generated from message com.hearlers.v1.service.UpdateContextResponse
- */
-export type UpdateContextResponse = Message<"com.hearlers.v1.service.UpdateContextResponse"> & {
-  /**
-   * @generated from field: com.hearlers.v1.model.Context context = 1;
-   */
-  context?: Context;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.UpdateContextResponse.
- * Use `create(UpdateContextResponseSchema)` to create a new message.
- */
-export const UpdateContextResponseSchema: GenMessage<UpdateContextResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 47);
-
-/**
- * Instruction
- *
- * @generated from message com.hearlers.v1.service.CreateInstructionRequest
- */
-export type CreateInstructionRequest = Message<"com.hearlers.v1.service.CreateInstructionRequest"> & {
-  /**
-   * @generated from field: string name = 1;
-   */
-  name: string;
-
-  /**
-   * @generated from field: optional string initial_sentence = 2;
-   */
-  initialSentence?: string;
-
-  /**
-   * @generated from field: repeated string instruction_item_ids = 3;
-   */
-  instructionItemIds: string[];
-};
-
-/**
- * Describes the message com.hearlers.v1.service.CreateInstructionRequest.
- * Use `create(CreateInstructionRequestSchema)` to create a new message.
- */
-export const CreateInstructionRequestSchema: GenMessage<CreateInstructionRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 48);
-
-/**
- * @generated from message com.hearlers.v1.service.CreateInstructionResponse
- */
-export type CreateInstructionResponse = Message<"com.hearlers.v1.service.CreateInstructionResponse"> & {
-  /**
-   * @generated from field: com.hearlers.v1.model.Instruction instruction = 1;
-   */
-  instruction?: Instruction;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.CreateInstructionResponse.
- * Use `create(CreateInstructionResponseSchema)` to create a new message.
- */
-export const CreateInstructionResponseSchema: GenMessage<CreateInstructionResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 49);
-
-/**
- * @generated from message com.hearlers.v1.service.FindInstructionsRequest
- */
-export type FindInstructionsRequest = Message<"com.hearlers.v1.service.FindInstructionsRequest"> & {
-  /**
-   * @generated from field: optional string name = 1;
-   */
-  name?: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindInstructionsRequest.
- * Use `create(FindInstructionsRequestSchema)` to create a new message.
- */
-export const FindInstructionsRequestSchema: GenMessage<FindInstructionsRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 50);
-
-/**
- * @generated from message com.hearlers.v1.service.FindInstructionsResponse
- */
-export type FindInstructionsResponse = Message<"com.hearlers.v1.service.FindInstructionsResponse"> & {
-  /**
-   * @generated from field: repeated com.hearlers.v1.model.Instruction instructions = 1;
-   */
-  instructions: Instruction[];
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindInstructionsResponse.
- * Use `create(FindInstructionsResponseSchema)` to create a new message.
- */
-export const FindInstructionsResponseSchema: GenMessage<FindInstructionsResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 51);
-
-/**
- * @generated from message com.hearlers.v1.service.FindInstructionByIdRequest
- */
-export type FindInstructionByIdRequest = Message<"com.hearlers.v1.service.FindInstructionByIdRequest"> & {
-  /**
-   * @generated from field: string instruction_id = 1;
-   */
-  instructionId: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindInstructionByIdRequest.
- * Use `create(FindInstructionByIdRequestSchema)` to create a new message.
- */
-export const FindInstructionByIdRequestSchema: GenMessage<FindInstructionByIdRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 52);
-
-/**
- * @generated from message com.hearlers.v1.service.FindInstructionByIdResponse
- */
-export type FindInstructionByIdResponse = Message<"com.hearlers.v1.service.FindInstructionByIdResponse"> & {
-  /**
-   * @generated from field: optional com.hearlers.v1.model.Instruction instruction = 1;
-   */
-  instruction?: Instruction;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindInstructionByIdResponse.
- * Use `create(FindInstructionByIdResponseSchema)` to create a new message.
- */
-export const FindInstructionByIdResponseSchema: GenMessage<FindInstructionByIdResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 53);
-
-/**
- * initial sentence 의 경우 nullable field
- * initial sentence 의 변경이 없을 경우 has_initial_sentence = false
- * initial sentence 를 null 로 변경할 경우 has_initial_sentence = true, initial_sentence = null
- * initial sentence 를 다른 값으로 변경할 경우 has_initial_sentence = true, initial_sentence = "new_initial_sentence"
- *
- * @generated from message com.hearlers.v1.service.UpdateInstructionRequest
- */
-export type UpdateInstructionRequest = Message<"com.hearlers.v1.service.UpdateInstructionRequest"> & {
-  /**
-   * @generated from field: string instruction_id = 1;
-   */
-  instructionId: string;
-
-  /**
-   * @generated from field: optional string name = 2;
-   */
-  name?: string;
-
-  /**
-   * @generated from field: optional string initial_sentence = 3;
-   */
-  initialSentence?: string;
-
-  /**
-   * @generated from field: bool has_initial_sentence = 4;
-   */
-  hasInitialSentence: boolean;
-
-  /**
-   * @generated from field: repeated string instruction_item_ids = 5;
-   */
-  instructionItemIds: string[];
-
-  /**
-   * @generated from field: bool has_instruction_item_ids = 6;
-   */
-  hasInstructionItemIds: boolean;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.UpdateInstructionRequest.
- * Use `create(UpdateInstructionRequestSchema)` to create a new message.
- */
-export const UpdateInstructionRequestSchema: GenMessage<UpdateInstructionRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 54);
-
-/**
- * @generated from message com.hearlers.v1.service.UpdateInstructionResponse
- */
-export type UpdateInstructionResponse = Message<"com.hearlers.v1.service.UpdateInstructionResponse"> & {
-  /**
-   * @generated from field: com.hearlers.v1.model.Instruction instruction = 1;
-   */
-  instruction?: Instruction;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.UpdateInstructionResponse.
- * Use `create(UpdateInstructionResponseSchema)` to create a new message.
- */
-export const UpdateInstructionResponseSchema: GenMessage<UpdateInstructionResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 55);
-
-/**
- * InstructionItem
- *
- * @generated from message com.hearlers.v1.service.CreateInstructionItemRequest
- */
-export type CreateInstructionItemRequest = Message<"com.hearlers.v1.service.CreateInstructionItemRequest"> & {
-  /**
-   * @generated from field: string body = 1;
-   */
-  body: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.CreateInstructionItemRequest.
- * Use `create(CreateInstructionItemRequestSchema)` to create a new message.
- */
-export const CreateInstructionItemRequestSchema: GenMessage<CreateInstructionItemRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 56);
-
-/**
- * @generated from message com.hearlers.v1.service.CreateInstructionItemResponse
- */
-export type CreateInstructionItemResponse = Message<"com.hearlers.v1.service.CreateInstructionItemResponse"> & {
-  /**
-   * @generated from field: com.hearlers.v1.model.InstructionItem instruction_item = 1;
-   */
-  instructionItem?: InstructionItem;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.CreateInstructionItemResponse.
- * Use `create(CreateInstructionItemResponseSchema)` to create a new message.
- */
-export const CreateInstructionItemResponseSchema: GenMessage<CreateInstructionItemResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 57);
-
-/**
- * @generated from message com.hearlers.v1.service.FindInstructionItemsRequest
- */
-export type FindInstructionItemsRequest = Message<"com.hearlers.v1.service.FindInstructionItemsRequest"> & {
-  /**
-   * @generated from field: optional string keyword = 1;
-   */
-  keyword?: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindInstructionItemsRequest.
- * Use `create(FindInstructionItemsRequestSchema)` to create a new message.
- */
-export const FindInstructionItemsRequestSchema: GenMessage<FindInstructionItemsRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 58);
-
-/**
- * @generated from message com.hearlers.v1.service.FindInstructionItemsResponse
- */
-export type FindInstructionItemsResponse = Message<"com.hearlers.v1.service.FindInstructionItemsResponse"> & {
-  /**
-   * @generated from field: repeated com.hearlers.v1.model.InstructionItem instruction_items = 1;
-   */
-  instructionItems: InstructionItem[];
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindInstructionItemsResponse.
- * Use `create(FindInstructionItemsResponseSchema)` to create a new message.
- */
-export const FindInstructionItemsResponseSchema: GenMessage<FindInstructionItemsResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 59);
-
-/**
- * @generated from message com.hearlers.v1.service.FindInstructionItemByIdRequest
- */
-export type FindInstructionItemByIdRequest = Message<"com.hearlers.v1.service.FindInstructionItemByIdRequest"> & {
-  /**
-   * @generated from field: string instruction_item_id = 1;
-   */
-  instructionItemId: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindInstructionItemByIdRequest.
- * Use `create(FindInstructionItemByIdRequestSchema)` to create a new message.
- */
-export const FindInstructionItemByIdRequestSchema: GenMessage<FindInstructionItemByIdRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 60);
-
-/**
- * @generated from message com.hearlers.v1.service.FindInstructionItemByIdResponse
- */
-export type FindInstructionItemByIdResponse = Message<"com.hearlers.v1.service.FindInstructionItemByIdResponse"> & {
-  /**
-   * @generated from field: optional com.hearlers.v1.model.InstructionItem instruction_item = 1;
-   */
-  instructionItem?: InstructionItem;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindInstructionItemByIdResponse.
- * Use `create(FindInstructionItemByIdResponseSchema)` to create a new message.
- */
-export const FindInstructionItemByIdResponseSchema: GenMessage<FindInstructionItemByIdResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 61);
-
-/**
- * @generated from message com.hearlers.v1.service.UpdateInstructionItemRequest
- */
-export type UpdateInstructionItemRequest = Message<"com.hearlers.v1.service.UpdateInstructionItemRequest"> & {
-  /**
-   * @generated from field: string instruction_item_id = 1;
-   */
-  instructionItemId: string;
-
-  /**
-   * @generated from field: optional string body = 2;
-   */
-  body?: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.UpdateInstructionItemRequest.
- * Use `create(UpdateInstructionItemRequestSchema)` to create a new message.
- */
-export const UpdateInstructionItemRequestSchema: GenMessage<UpdateInstructionItemRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 62);
-
-/**
- * @generated from message com.hearlers.v1.service.UpdateInstructionItemResponse
- */
-export type UpdateInstructionItemResponse = Message<"com.hearlers.v1.service.UpdateInstructionItemResponse"> & {
-  /**
-   * @generated from field: com.hearlers.v1.model.InstructionItem instruction_item = 1;
-   */
-  instructionItem?: InstructionItem;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.UpdateInstructionItemResponse.
- * Use `create(UpdateInstructionItemResponseSchema)` to create a new message.
- */
-export const UpdateInstructionItemResponseSchema: GenMessage<UpdateInstructionItemResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 63);
-
-/**
- * Tone
- *
- * @generated from message com.hearlers.v1.service.CreateToneRequest
- */
-export type CreateToneRequest = Message<"com.hearlers.v1.service.CreateToneRequest"> & {
-  /**
-   * @generated from field: string name = 1;
-   */
-  name: string;
-
-  /**
-   * @generated from field: string body = 2;
-   */
-  body: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.CreateToneRequest.
- * Use `create(CreateToneRequestSchema)` to create a new message.
- */
-export const CreateToneRequestSchema: GenMessage<CreateToneRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 64);
-
-/**
- * @generated from message com.hearlers.v1.service.CreateToneResponse
- */
-export type CreateToneResponse = Message<"com.hearlers.v1.service.CreateToneResponse"> & {
-  /**
-   * @generated from field: com.hearlers.v1.model.Tone tone = 1;
-   */
-  tone?: Tone;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.CreateToneResponse.
- * Use `create(CreateToneResponseSchema)` to create a new message.
- */
-export const CreateToneResponseSchema: GenMessage<CreateToneResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 65);
-
-/**
- * @generated from message com.hearlers.v1.service.FindTonesRequest
- */
-export type FindTonesRequest = Message<"com.hearlers.v1.service.FindTonesRequest"> & {
-  /**
-   * @generated from field: optional string name = 1;
-   */
-  name?: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindTonesRequest.
- * Use `create(FindTonesRequestSchema)` to create a new message.
- */
-export const FindTonesRequestSchema: GenMessage<FindTonesRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 66);
-
-/**
- * @generated from message com.hearlers.v1.service.FindTonesResponse
- */
-export type FindTonesResponse = Message<"com.hearlers.v1.service.FindTonesResponse"> & {
-  /**
-   * @generated from field: repeated com.hearlers.v1.model.Tone tones = 1;
-   */
-  tones: Tone[];
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindTonesResponse.
- * Use `create(FindTonesResponseSchema)` to create a new message.
- */
-export const FindTonesResponseSchema: GenMessage<FindTonesResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 67);
-
-/**
- * @generated from message com.hearlers.v1.service.FindToneByIdRequest
- */
-export type FindToneByIdRequest = Message<"com.hearlers.v1.service.FindToneByIdRequest"> & {
-  /**
-   * @generated from field: string tone_id = 1;
-   */
-  toneId: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindToneByIdRequest.
- * Use `create(FindToneByIdRequestSchema)` to create a new message.
- */
-export const FindToneByIdRequestSchema: GenMessage<FindToneByIdRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 68);
-
-/**
- * @generated from message com.hearlers.v1.service.FindToneByIdResponse
- */
-export type FindToneByIdResponse = Message<"com.hearlers.v1.service.FindToneByIdResponse"> & {
-  /**
-   * @generated from field: optional com.hearlers.v1.model.Tone tone = 1;
-   */
-  tone?: Tone;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.FindToneByIdResponse.
- * Use `create(FindToneByIdResponseSchema)` to create a new message.
- */
-export const FindToneByIdResponseSchema: GenMessage<FindToneByIdResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 69);
-
-/**
- * @generated from message com.hearlers.v1.service.UpdateToneRequest
- */
-export type UpdateToneRequest = Message<"com.hearlers.v1.service.UpdateToneRequest"> & {
-  /**
-   * @generated from field: string tone_id = 1;
-   */
-  toneId: string;
-
-  /**
-   * @generated from field: optional string name = 2;
-   */
-  name?: string;
-
-  /**
-   * @generated from field: optional string body = 3;
-   */
-  body?: string;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.UpdateToneRequest.
- * Use `create(UpdateToneRequestSchema)` to create a new message.
- */
-export const UpdateToneRequestSchema: GenMessage<UpdateToneRequest> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 70);
-
-/**
- * @generated from message com.hearlers.v1.service.UpdateToneResponse
- */
-export type UpdateToneResponse = Message<"com.hearlers.v1.service.UpdateToneResponse"> & {
-  /**
-   * @generated from field: com.hearlers.v1.model.Tone tone = 1;
-   */
-  tone?: Tone;
-};
-
-/**
- * Describes the message com.hearlers.v1.service.UpdateToneResponse.
- * Use `create(UpdateToneResponseSchema)` to create a new message.
- */
-export const UpdateToneResponseSchema: GenMessage<UpdateToneResponse> = /*@__PURE__*/
-  messageDesc(file_com_hearlers_v1_service_counsel, 71);
+  messageDesc(file_com_hearlers_v1_service_counsel, 13);
 
 /**
  * @generated from service com.hearlers.v1.service.CounselService
@@ -1519,8 +319,6 @@ export const CounselService: GenService<{
     output: typeof CreateCounselResponseSchema;
   },
   /**
-   * TODO: 상담사 별로 조회
-   *
    * @generated from rpc com.hearlers.v1.service.CounselService.FindCounsels
    */
   findCounsels: {
@@ -1537,7 +335,7 @@ export const CounselService: GenService<{
     output: typeof FindCounselByIdResponseSchema;
   },
   /**
-   * 상담 메시지 (상담 중 하나의 채팅팅)
+   * 상담 메시지 (상담 중 하나의 채팅)
    *
    * @generated from rpc com.hearlers.v1.service.CounselService.CreateMessage
    */
@@ -1563,40 +361,6 @@ export const CounselService: GenService<{
     output: typeof ReactMessageResponseSchema;
   },
   /**
-   * 상담사
-   *
-   * @generated from rpc com.hearlers.v1.service.CounselService.CreateCounselor
-   */
-  createCounselor: {
-    methodKind: "unary";
-    input: typeof CreateCounselorRequestSchema;
-    output: typeof CreateCounselorResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.FindCounselors
-   */
-  findCounselors: {
-    methodKind: "unary";
-    input: typeof FindCounselorsRequestSchema;
-    output: typeof FindCounselorsResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.FindCounselorById
-   */
-  findCounselorById: {
-    methodKind: "unary";
-    input: typeof FindCounselorByIdRequestSchema;
-    output: typeof FindCounselorByIdResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.UpdateCounselor
-   */
-  updateCounselor: {
-    methodKind: "unary";
-    input: typeof UpdateCounselorRequestSchema;
-    output: typeof UpdateCounselorResponseSchema;
-  },
-  /**
    * 상담사와 유저의 관계
    * TODO: CQRS 패턴으로 뷰 테이블을 만들어 놓고 그냥 꺼내서 준다.
    * Relationship은 읽을 시 싱딤시와, 상담사의 모든 Counsels를 다 가지고 오는게 의미있는 단위
@@ -1610,218 +374,6 @@ export const CounselService: GenService<{
     methodKind: "unary";
     input: typeof FindCounselorUserRelationshipsRequestSchema;
     output: typeof FindCounselorUserRelationshipsResponseSchema;
-  },
-  /**
-   * 상담 기법
-   *
-   * @generated from rpc com.hearlers.v1.service.CounselService.CreateCounselTechnique
-   */
-  createCounselTechnique: {
-    methodKind: "unary";
-    input: typeof CreateCounselTechniqueRequestSchema;
-    output: typeof CreateCounselTechniqueResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.FindCounselTechniques
-   */
-  findCounselTechniques: {
-    methodKind: "unary";
-    input: typeof FindCounselTechniquesRequestSchema;
-    output: typeof FindCounselTechniquesResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.FindCounselTechniqueById
-   */
-  findCounselTechniqueById: {
-    methodKind: "unary";
-    input: typeof FindCounselTechniqueByIdRequestSchema;
-    output: typeof FindCounselTechniqueByIdResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.UpdateCounselTechnique
-   */
-  updateCounselTechnique: {
-    methodKind: "unary";
-    input: typeof UpdateCounselTechniqueRequestSchema;
-    output: typeof UpdateCounselTechniqueResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.SaveCounselTechniqueSequence
-   */
-  saveCounselTechniqueSequence: {
-    methodKind: "unary";
-    input: typeof SaveCounselTechniqueSequenceRequestSchema;
-    output: typeof SaveCounselTechniqueSequenceResponseSchema;
-  },
-  /**
-   * Persona
-   *
-   * @generated from rpc com.hearlers.v1.service.CounselService.CreatePersona
-   */
-  createPersona: {
-    methodKind: "unary";
-    input: typeof CreatePersonaRequestSchema;
-    output: typeof CreatePersonaResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.FindPersonas
-   */
-  findPersonas: {
-    methodKind: "unary";
-    input: typeof FindPersonasRequestSchema;
-    output: typeof FindPersonasResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.FindPersonaById
-   */
-  findPersonaById: {
-    methodKind: "unary";
-    input: typeof FindPersonaByIdRequestSchema;
-    output: typeof FindPersonaByIdResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.UpdatePersona
-   */
-  updatePersona: {
-    methodKind: "unary";
-    input: typeof UpdatePersonaRequestSchema;
-    output: typeof UpdatePersonaResponseSchema;
-  },
-  /**
-   * Context
-   *
-   * @generated from rpc com.hearlers.v1.service.CounselService.CreateContext
-   */
-  createContext: {
-    methodKind: "unary";
-    input: typeof CreateContextRequestSchema;
-    output: typeof CreateContextResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.FindContexts
-   */
-  findContexts: {
-    methodKind: "unary";
-    input: typeof FindContextsRequestSchema;
-    output: typeof FindContextsResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.FindContextById
-   */
-  findContextById: {
-    methodKind: "unary";
-    input: typeof FindContextByIdRequestSchema;
-    output: typeof FindContextByIdResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.UpdateContext
-   */
-  updateContext: {
-    methodKind: "unary";
-    input: typeof UpdateContextRequestSchema;
-    output: typeof UpdateContextResponseSchema;
-  },
-  /**
-   * Instruction
-   *
-   * @generated from rpc com.hearlers.v1.service.CounselService.CreateInstruction
-   */
-  createInstruction: {
-    methodKind: "unary";
-    input: typeof CreateInstructionRequestSchema;
-    output: typeof CreateInstructionResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.FindInstructions
-   */
-  findInstructions: {
-    methodKind: "unary";
-    input: typeof FindInstructionsRequestSchema;
-    output: typeof FindInstructionsResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.FindInstructionById
-   */
-  findInstructionById: {
-    methodKind: "unary";
-    input: typeof FindInstructionByIdRequestSchema;
-    output: typeof FindInstructionByIdResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.UpdateInstruction
-   */
-  updateInstruction: {
-    methodKind: "unary";
-    input: typeof UpdateInstructionRequestSchema;
-    output: typeof UpdateInstructionResponseSchema;
-  },
-  /**
-   * InstructionItem
-   *
-   * @generated from rpc com.hearlers.v1.service.CounselService.CreateInstructionItem
-   */
-  createInstructionItem: {
-    methodKind: "unary";
-    input: typeof CreateInstructionItemRequestSchema;
-    output: typeof CreateInstructionItemResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.FindInstructionItems
-   */
-  findInstructionItems: {
-    methodKind: "unary";
-    input: typeof FindInstructionItemsRequestSchema;
-    output: typeof FindInstructionItemsResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.FindInstructionItemById
-   */
-  findInstructionItemById: {
-    methodKind: "unary";
-    input: typeof FindInstructionItemByIdRequestSchema;
-    output: typeof FindInstructionItemByIdResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.UpdateInstructionItem
-   */
-  updateInstructionItem: {
-    methodKind: "unary";
-    input: typeof UpdateInstructionItemRequestSchema;
-    output: typeof UpdateInstructionItemResponseSchema;
-  },
-  /**
-   * Tone
-   *
-   * @generated from rpc com.hearlers.v1.service.CounselService.CreateTone
-   */
-  createTone: {
-    methodKind: "unary";
-    input: typeof CreateToneRequestSchema;
-    output: typeof CreateToneResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.FindTones
-   */
-  findTones: {
-    methodKind: "unary";
-    input: typeof FindTonesRequestSchema;
-    output: typeof FindTonesResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.FindToneById
-   */
-  findToneById: {
-    methodKind: "unary";
-    input: typeof FindToneByIdRequestSchema;
-    output: typeof FindToneByIdResponseSchema;
-  },
-  /**
-   * @generated from rpc com.hearlers.v1.service.CounselService.UpdateTone
-   */
-  updateTone: {
-    methodKind: "unary";
-    input: typeof UpdateToneRequestSchema;
-    output: typeof UpdateToneResponseSchema;
   },
 }> = /*@__PURE__*/
   serviceDesc(file_com_hearlers_v1_service_counsel, 0);
