+++ src/gen/com/hearlers/v1/service/counselor_pb.ts
@@ -1,0 +1,240 @@
+// @generated by protoc-gen-es v2.2.1 with parameter "target=ts"
+// @generated from file com/hearlers/v1/service/counselor.proto (package com.hearlers.v1.service, syntax proto3)
+/* eslint-disable */
+
+import type { GenFile, GenMessage, GenService } from "@bufbuild/protobuf/codegenv1";
+import { fileDesc, messageDesc, serviceDesc } from "@bufbuild/protobuf/codegenv1";
+import type { Counselor, CounselorGender } from "../model/counselor_pb";
+import { file_com_hearlers_v1_model_counselor } from "../model/counselor_pb";
+import type { Message } from "@bufbuild/protobuf";
+
+/**
+ * Describes the file com/hearlers/v1/service/counselor.proto.
+ */
+export const file_com_hearlers_v1_service_counselor: GenFile = /*@__PURE__*/
+  fileDesc("Cidjb20vaGVhcmxlcnMvdjEvc2VydmljZS9jb3Vuc2Vsb3IucHJvdG8SF2NvbS5oZWFybGVycy52MS5zZXJ2aWNlIp8BChZDcmVhdGVDb3Vuc2Vsb3JSZXF1ZXN0Eg8KB3RvbmVfaWQYASABKAkSDAoEbmFtZRgCIAEoCRITCgtkZXNjcmlwdGlvbhgDIAEoCRJAChBjb3Vuc2Vsb3JfZ2VuZGVyGAQgASgOMiYuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxvckdlbmRlchIPCgdwZXJzb25hGAUgASgJIk4KF0NyZWF0ZUNvdW5zZWxvclJlc3BvbnNlEjMKCWNvdW5zZWxvchgBIAEoCzIgLmNvbS5oZWFybGVycy52MS5tb2RlbC5Db3Vuc2Vsb3IiOQoVRmluZENvdW5zZWxvcnNSZXF1ZXN0EhQKB3RvbmVfaWQYASABKAlIAIgBAUIKCghfdG9uZV9pZCJOChZGaW5kQ291bnNlbG9yc1Jlc3BvbnNlEjQKCmNvdW5zZWxvcnMYASADKAsyIC5jb20uaGVhcmxlcnMudjEubW9kZWwuQ291bnNlbG9yIjAKGEZpbmRDb3Vuc2Vsb3JCeUlkUmVxdWVzdBIUCgxjb3Vuc2Vsb3JfaWQYASABKAkiYwoZRmluZENvdW5zZWxvckJ5SWRSZXNwb25zZRI4Cgljb3Vuc2Vsb3IYASABKAsyIC5jb20uaGVhcmxlcnMudjEubW9kZWwuQ291bnNlbG9ySACIAQFCDAoKX2NvdW5zZWxvciKUAgoWVXBkYXRlQ291bnNlbG9yUmVxdWVzdBIUCgxjb3Vuc2Vsb3JfaWQYASABKAkSFAoHdG9uZV9pZBgCIAEoCUgAiAEBEhEKBG5hbWUYAyABKAlIAYgBARIYCgtkZXNjcmlwdGlvbhgEIAEoCUgCiAEBEkUKEGNvdW5zZWxvcl9nZW5kZXIYBSABKA4yJi5jb20uaGVhcmxlcnMudjEubW9kZWwuQ291bnNlbG9yR2VuZGVySAOIAQESFAoHcGVyc29uYRgGIAEoCUgEiAEBQgoKCF90b25lX2lkQgcKBV9uYW1lQg4KDF9kZXNjcmlwdGlvbkITChFfY291bnNlbG9yX2dlbmRlckIKCghfcGVyc29uYSJOChdVcGRhdGVDb3Vuc2Vsb3JSZXNwb25zZRIzCgljb3Vuc2Vsb3IYASABKAsyIC5jb20uaGVhcmxlcnMudjEubW9kZWwuQ291bnNlbG9yMu0DChBDb3Vuc2Vsb3JTZXJ2aWNlEnQKD0NyZWF0ZUNvdW5zZWxvchIvLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkNyZWF0ZUNvdW5zZWxvclJlcXVlc3QaMC5jb20uaGVhcmxlcnMudjEuc2VydmljZS5DcmVhdGVDb3Vuc2Vsb3JSZXNwb25zZRJxCg5GaW5kQ291bnNlbG9ycxIuLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkZpbmRDb3Vuc2Vsb3JzUmVxdWVzdBovLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkZpbmRDb3Vuc2Vsb3JzUmVzcG9uc2USegoRRmluZENvdW5zZWxvckJ5SWQSMS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kQ291bnNlbG9yQnlJZFJlcXVlc3QaMi5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kQ291bnNlbG9yQnlJZFJlc3BvbnNlEnQKD1VwZGF0ZUNvdW5zZWxvchIvLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLlVwZGF0ZUNvdW5zZWxvclJlcXVlc3QaMC5jb20uaGVhcmxlcnMudjEuc2VydmljZS5VcGRhdGVDb3Vuc2Vsb3JSZXNwb25zZUKtAQobY29tLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlQg5Db3Vuc2Vsb3JQcm90b1ABogIEQ0hWU6oCF0NvbS5IZWFybGVycy5WMS5TZXJ2aWNlygIXQ29tXEhlYXJsZXJzXFYxXFNlcnZpY2XiAiNDb21cSGVhcmxlcnNcVjFcU2VydmljZVxHUEJNZXRhZGF0YeoCGkNvbTo6SGVhcmxlcnM6OlYxOjpTZXJ2aWNlYgZwcm90bzM", [file_com_hearlers_v1_model_counselor]);
+
+/**
+ * 상담사
+ *
+ * @generated from message com.hearlers.v1.service.CreateCounselorRequest
+ */
+export type CreateCounselorRequest = Message<"com.hearlers.v1.service.CreateCounselorRequest"> & {
+  /**
+   * @generated from field: string tone_id = 1;
+   */
+  toneId: string;
+
+  /**
+   * @generated from field: string name = 2;
+   */
+  name: string;
+
+  /**
+   * @generated from field: string description = 3;
+   */
+  description: string;
+
+  /**
+   * @generated from field: com.hearlers.v1.model.CounselorGender counselor_gender = 4;
+   */
+  counselorGender: CounselorGender;
+
+  /**
+   * @generated from field: string persona = 5;
+   */
+  persona: string;
+};
+
+/**
+ * Describes the message com.hearlers.v1.service.CreateCounselorRequest.
+ * Use `create(CreateCounselorRequestSchema)` to create a new message.
+ */
+export const CreateCounselorRequestSchema: GenMessage<CreateCounselorRequest> = /*@__PURE__*/
+  messageDesc(file_com_hearlers_v1_service_counselor, 0);
+
+/**
+ * @generated from message com.hearlers.v1.service.CreateCounselorResponse
+ */
+export type CreateCounselorResponse = Message<"com.hearlers.v1.service.CreateCounselorResponse"> & {
+  /**
+   * @generated from field: com.hearlers.v1.model.Counselor counselor = 1;
+   */
+  counselor?: Counselor;
+};
+
+/**
+ * Describes the message com.hearlers.v1.service.CreateCounselorResponse.
+ * Use `create(CreateCounselorResponseSchema)` to create a new message.
+ */
+export const CreateCounselorResponseSchema: GenMessage<CreateCounselorResponse> = /*@__PURE__*/
+  messageDesc(file_com_hearlers_v1_service_counselor, 1);
+
+/**
+ * @generated from message com.hearlers.v1.service.FindCounselorsRequest
+ */
+export type FindCounselorsRequest = Message<"com.hearlers.v1.service.FindCounselorsRequest"> & {
+  /**
+   * @generated from field: optional string tone_id = 1;
+   */
+  toneId?: string;
+};
+
+/**
+ * Describes the message com.hearlers.v1.service.FindCounselorsRequest.
+ * Use `create(FindCounselorsRequestSchema)` to create a new message.
+ */
+export const FindCounselorsRequestSchema: GenMessage<FindCounselorsRequest> = /*@__PURE__*/
+  messageDesc(file_com_hearlers_v1_service_counselor, 2);
+
+/**
+ * @generated from message com.hearlers.v1.service.FindCounselorsResponse
+ */
+export type FindCounselorsResponse = Message<"com.hearlers.v1.service.FindCounselorsResponse"> & {
+  /**
+   * @generated from field: repeated com.hearlers.v1.model.Counselor counselors = 1;
+   */
+  counselors: Counselor[];
+};
+
+/**
+ * Describes the message com.hearlers.v1.service.FindCounselorsResponse.
+ * Use `create(FindCounselorsResponseSchema)` to create a new message.
+ */
+export const FindCounselorsResponseSchema: GenMessage<FindCounselorsResponse> = /*@__PURE__*/
+  messageDesc(file_com_hearlers_v1_service_counselor, 3);
+
+/**
+ * @generated from message com.hearlers.v1.service.FindCounselorByIdRequest
+ */
+export type FindCounselorByIdRequest = Message<"com.hearlers.v1.service.FindCounselorByIdRequest"> & {
+  /**
+   * @generated from field: string counselor_id = 1;
+   */
+  counselorId: string;
+};
+
+/**
+ * Describes the message com.hearlers.v1.service.FindCounselorByIdRequest.
+ * Use `create(FindCounselorByIdRequestSchema)` to create a new message.
+ */
+export const FindCounselorByIdRequestSchema: GenMessage<FindCounselorByIdRequest> = /*@__PURE__*/
+  messageDesc(file_com_hearlers_v1_service_counselor, 4);
+
+/**
+ * @generated from message com.hearlers.v1.service.FindCounselorByIdResponse
+ */
+export type FindCounselorByIdResponse = Message<"com.hearlers.v1.service.FindCounselorByIdResponse"> & {
+  /**
+   * @generated from field: optional com.hearlers.v1.model.Counselor counselor = 1;
+   */
+  counselor?: Counselor;
+};
+
+/**
+ * Describes the message com.hearlers.v1.service.FindCounselorByIdResponse.
+ * Use `create(FindCounselorByIdResponseSchema)` to create a new message.
+ */
+export const FindCounselorByIdResponseSchema: GenMessage<FindCounselorByIdResponse> = /*@__PURE__*/
+  messageDesc(file_com_hearlers_v1_service_counselor, 5);
+
+/**
+ * @generated from message com.hearlers.v1.service.UpdateCounselorRequest
+ */
+export type UpdateCounselorRequest = Message<"com.hearlers.v1.service.UpdateCounselorRequest"> & {
+  /**
+   * @generated from field: string counselor_id = 1;
+   */
+  counselorId: string;
+
+  /**
+   * @generated from field: optional string tone_id = 2;
+   */
+  toneId?: string;
+
+  /**
+   * @generated from field: optional string name = 3;
+   */
+  name?: string;
+
+  /**
+   * @generated from field: optional string description = 4;
+   */
+  description?: string;
+
+  /**
+   * @generated from field: optional com.hearlers.v1.model.CounselorGender counselor_gender = 5;
+   */
+  counselorGender?: CounselorGender;
+
+  /**
+   * @generated from field: optional string persona = 6;
+   */
+  persona?: string;
+};
+
+/**
+ * Describes the message com.hearlers.v1.service.UpdateCounselorRequest.
+ * Use `create(UpdateCounselorRequestSchema)` to create a new message.
+ */
+export const UpdateCounselorRequestSchema: GenMessage<UpdateCounselorRequest> = /*@__PURE__*/
+  messageDesc(file_com_hearlers_v1_service_counselor, 6);
+
+/**
+ * @generated from message com.hearlers.v1.service.UpdateCounselorResponse
+ */
+export type UpdateCounselorResponse = Message<"com.hearlers.v1.service.UpdateCounselorResponse"> & {
+  /**
+   * @generated from field: com.hearlers.v1.model.Counselor counselor = 1;
+   */
+  counselor?: Counselor;
+};
+
+/**
+ * Describes the message com.hearlers.v1.service.UpdateCounselorResponse.
+ * Use `create(UpdateCounselorResponseSchema)` to create a new message.
+ */
+export const UpdateCounselorResponseSchema: GenMessage<UpdateCounselorResponse> = /*@__PURE__*/
+  messageDesc(file_com_hearlers_v1_service_counselor, 7);
+
+/**
+ * @generated from service com.hearlers.v1.service.CounselorService
+ */
+export const CounselorService: GenService<{
+  /**
+   * 상담사
+   *
+   * @generated from rpc com.hearlers.v1.service.CounselorService.CreateCounselor
+   */
+  createCounselor: {
+    methodKind: "unary";
+    input: typeof CreateCounselorRequestSchema;
+    output: typeof CreateCounselorResponseSchema;
+  },
+  /**
+   * @generated from rpc com.hearlers.v1.service.CounselorService.FindCounselors
+   */
+  findCounselors: {
+    methodKind: "unary";
+    input: typeof FindCounselorsRequestSchema;
+    output: typeof FindCounselorsResponseSchema;
+  },
+  /**
+   * @generated from rpc com.hearlers.v1.service.CounselorService.FindCounselorById
+   */
+  findCounselorById: {
+    methodKind: "unary";
+    input: typeof FindCounselorByIdRequestSchema;
+    output: typeof FindCounselorByIdResponseSchema;
+  },
+  /**
+   * @generated from rpc com.hearlers.v1.service.CounselorService.UpdateCounselor
+   */
+  updateCounselor: {
+    methodKind: "unary";
+    input: typeof UpdateCounselorRequestSchema;
+    output: typeof UpdateCounselorResponseSchema;
+  },
+}> = /*@__PURE__*/
+  serviceDesc(file_com_hearlers_v1_service_counselor, 0);
+
